### 2020-06-27 69a92c1
cleaned the code and added some comments
