### 2020-06-27 #2398 @florian-huber
Reviewers from the list that could fit both topic-wise (Cheminformatics, mass spectroscopy) and language-wise (Python):

- hannahbrucemacdonald
- abotzki
- mberacochea
- thesketh
- RMeli
- bovee

Further suggestions (not from the list):
- lfnothias
- mwang87
