### 2020-06-27 6a83ef4
add workflow figure to index.rst
