### 2020-06-27 89dc35f
Switch to newer workflow figure
