### 2020-06-27 e620b79
add example plot to documentation
