+++ matchms/Spectrum.py
@@ -33,6 +33,14 @@ class Spectrum:
         0.7
         spectrum1
 
+    To visually inspect a spectrum run ``spectrum.plot()``
+
+    .. figure:: paper/spectrum-plot-example.png
+        :width: 400
+        :alt: spectrum plotting function
+
+        Example of a spectrum plotted using ``spectrum.plot()`` and ``spectrum.plot(intensity_to=0.02)``.
+
     Attributes
     ----------
     peaks: ~matchms.Spikes.Spikes
