### 2020-06-29 48c8b1c
Update readthedocs/index.rst
Co-authored-by: Stefan Verhoeven <s.verhoeven@esciencecenter.nl>
