### 2020-06-29 dab9bd7
move example images to readthedocs/_static
