+++ readthedocs/index.rst
@@ -21,8 +21,8 @@ Introduction
 
 Matchms allows to easily build custom spectra processing pipelines and to compute spectra similarities.
 
-.. image:: ../paper/flowchart_matchms.png
-  :width: 500
+.. image:: _static/flowchart_matchms.png
+  :width: 400
   :alt: matchms workflow illustration
 
 Example
