### 2020-06-29 0753036
Update matchms/Spectrum.py
Co-authored-by: Stefan Verhoeven <s.verhoeven@esciencecenter.nl>
