### 2020-06-29 5591192
change static folder link and remove whitespaces
