+++ matchms/Spectrum.py
@@ -90,7 +90,7 @@ class Spectrum:
         .. figure:: ../_static/spectrum-plot-example.png
             :width: 400
             :alt: spectrum plotting function
-    
+
             Example of a spectrum plotted using ``spectrum.plot()`` and ``spectrum.plot(intensity_to=0.02)``.."""
 
         def plot_histogram():
