### 2020-06-29 b02f5d8
Update readthedocs/index.rst
Co-authored-by: Stefan Verhoeven <s.verhoeven@esciencecenter.nl>
