### 2020-06-29 96f8eff
CosineGreedyNumba -> CosineGreedy and former CosineGreedy -> CosineGreedyVectorial
