+++ matchms/similarity/CosineGreedy.py
@@ -1,17 +1,18 @@
 from typing import Tuple
 import numpy
+from matchms.similarity.collect_peak_pairs import collect_peak_pairs
 from matchms.typing import SpectrumType
 
 
 class CosineGreedy:
-    """Factory to calculate 'greedy cosine score' between mass spectra.
+    """Calculate 'cosine similarity score' between two spectra.
 
     The cosine score aims at quantifying the similarity between two mass spectra.
     The score is calculated by finding best possible matches between peaks
     of two spectra. Two peaks are considered a potential match if their
     m/z ratios lie within the given 'tolerance'.
-    The underlying peak assignment problem is here solved in a 'greedy' way.
-    This can perform notably faster, but does occasionally deviate slightly from
+    The underlying peak assignment problem is here sovled in a 'greedy' way.
+    This can perform notably faster, but does occationally deviate slightly from
     a fully correct solution (as with the Hungarian algorithm). In practice this
     will rarely affect similarity scores notably, in particular for smaller
     tolerances.
@@ -43,81 +44,56 @@ class CosineGreedy:
         Cosine score is 0.52 with 1 matched peaks
 
     """
-    def __init__(self, tolerance: float = 0.3):
+    def __init__(self, tolerance=0.1):
         """
-
-        Parameters
-        ----------
-        tolerance
-            Peaks will be considered a match when <= tolerance apart.
+        Args:
+        ----
+        tolerance: float
+            Peaks will be considered a match when <= tolerance appart. Default is 0.1.
         """
         self.tolerance = tolerance
 
-    def __call__(self, spectrum: SpectrumType, reference_spectrum: SpectrumType) -> Tuple[float, int]:
-        """Calculate 'greedy cosine score' between mass spectra.
-
+    def __call__(self, spectrum1: SpectrumType, spectrum2: SpectrumType) -> Tuple[float, int]:
+        """Calculate cosine score between two spectra.
         Args:
-        -----
-        spectrum
-            First spectrum
-        reference_spectrum
-            Second spectrum
-
-        Returns:
-        --------
-
-        Tuple with cosine score and number of matched peaks.
+        ----
+        spectrum1: SpectrumType
+            Input spectrum 1.
+        spectrum2: SpectrumType
+            Input spectrum 2.
         """
-        def calc_mz_distance():
-            mz_row_vector = spectrum.peaks.mz
-            mz_col_vector = numpy.reshape(reference_spectrum.peaks.mz, (n_rows, 1))
+        def get_peaks_arrays():
+            """Get peaks mz and intensities as numpy array."""
+            spec1 = numpy.vstack((spectrum1.peaks.mz, spectrum1.peaks.intensities)).T
+            spec2 = numpy.vstack((spectrum2.peaks.mz, spectrum2.peaks.intensities)).T
+            assert max(spec1[:, 1]) <= 1, ("Input spectrum1 is not normalized. ",
+                                           "Apply 'normalize_intensities' filter first.")
+            assert max(spec2[:, 1]) <= 1, ("Input spectrum2 is not normalized. ",
+                                           "Apply 'normalize_intensities' filter first.")
+            return spec1, spec2
 
-            mz1 = numpy.tile(mz_row_vector, (n_rows, 1))
-            mz2 = numpy.tile(mz_col_vector, (1, n_cols))
-
-            return mz1 - mz2
-
-        def calc_intensities_product():
-            intensities_row_vector = spectrum.peaks.intensities
-            intensities_col_vector = numpy.reshape(reference_spectrum.peaks.intensities, (n_rows, 1))
-
-            intensities1 = numpy.tile(intensities_row_vector, (n_rows, 1))
-            intensities2 = numpy.tile(intensities_col_vector, (1, n_cols))
-
[…]
