+++ matchms/similarity/__init__.py
@@ -1,6 +1,6 @@
 """similarity module"""
 from .CosineGreedy import CosineGreedy
-from .CosineGreedyNumba import CosineGreedyNumba
+from .CosineGreedyVectorial import CosineGreedyVectorial
 from .CosineHungarian import CosineHungarian
 from .FingerprintSimilarityParallel import FingerprintSimilarityParallel
 from .IntersectMz import IntersectMz
@@ -11,7 +11,7 @@ from .ParentmassMatchParallel import ParentmassMatchParallel
 
 __all__ = [
     "CosineGreedy",
-    "CosineGreedyNumba",
+    "CosineGreedyVectorial",
     "CosineHungarian",
     "FingerprintSimilarityParallel",
     "IntersectMz",

[…]
