### 2020-06-29 d673035
extend docstring
