+++ matchms/similarity/CosineGreedyVectorial.py
@@ -4,7 +4,7 @@ from matchms.typing import SpectrumType
 
 
 class CosineGreedyVectorial:
-    """Factory to calculate 'greedy cosine score' between mass spectra.
+    """Calculate 'cosine similarity score' between two spectra.
 
     The cosine score aims at quantifying the similarity between two mass spectra.
     The score is calculated by finding best possible matches between peaks
