### 2020-06-29 08685b8
minor corrections to docstring and types
