+++ matchms/similarity/CosineGreedyVectorial.py
@@ -43,7 +43,7 @@ class CosineGreedyVectorial:
         Cosine score is 0.52 with 1 matched peaks
 
     """
-    def __init__(self, tolerance: float = 0.3):
+    def __init__(self, tolerance: float = 0.1):
         """
 
         Parameters
