### 2020-06-29 38604ef
update index example to new tolerance default
