### 2020-06-29 ff03752
improved part of the reading file code
