### 2020-06-29 2d65b7b
updated the documentarion and changelog files
