+++ CHANGELOG.md
@@ -9,6 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Read msp files and created Spectrum objects from it
+
+## [0.4.1] - 2020-06-29
+
+### Added
+
 - Added similarity score based on comparing parent masses [#79](https://github.com/matchms/matchms/pull/79)
 
 ## [0.4.0] - 2020-06-11
