### 2020-06-29 d4b7b00
reduce code duplication and extend docstrings
