+++ matchms/similarity/ModifiedCosine.py
@@ -1,6 +1,7 @@
 from typing import Tuple
-import numpy
-from matchms.similarity.collect_peak_pairs import collect_peak_pairs
+from matchms.similarity.spectrum_similarity_functions import collect_peak_pairs
+from matchms.similarity.spectrum_similarity_functions import get_peaks_array
+from matchms.similarity.spectrum_similarity_functions import score_best_matches
 from matchms.typing import SpectrumType
 
 
@@ -16,12 +17,42 @@ class ModifiedCosine:
     See Watrous et al. [PNAS, 2012, https://www.pnas.org/content/109/26/E1743]
     for further details.
 
-    Args:
-    -----
-    tolerance: float
-        Peaks will be considered a match when <= tolerance a part. Default is 0.1.
+    For example
+
+    .. testcode::
+
+        import numpy as np
+        from matchms import Spectrum
+        from matchms.similarity import ModifiedCosine
+
+        spectrum_1 = Spectrum(mz=np.array([100, 150, 200.]),
+                              intensities=np.array([0.7, 0.2, 0.1]),
+                              metadata={"precursor_mz": 100.0})
+        spectrum_2 = Spectrum(mz=np.array([104.9, 140, 190.]),
+                              intensities=np.array([0.4, 0.2, 0.1]),
+                              metadata={"precursor_mz": 105.0})
+
+        # Use factory to construct a similarity function
+        modified_cosine = ModifiedCosine(tolerance=0.2)
+
+        score, n_matches = modified_cosine(spectrum_1, spectrum_2)
+
+        print(f"Modified cosine score is {score:.2f} with {n_matches} matched peaks")
+
+    Should output
+
+    .. testoutput::
+
+        Cosine score is 0.52 with 1 matched peaks
+
     """
-    def __init__(self, tolerance=0.1):
+    def __init__(self, tolerance: float = 0.1):
+        """
+        Args:
+        -----
+        tolerance
+            Peaks will be considered a match when <= tolerance a part. Default is 0.1.
+        """
         self.tolerance = tolerance
 
     def __call__(self, spectrum1: SpectrumType, spectrum2: SpectrumType) -> Tuple[float, int]:
@@ -38,16 +69,6 @@ class ModifiedCosine:
 
         Tuple with cosine score and number of matched peaks.
         """
-        def get_peaks_arrays():
-            """Get peaks mz and intensities as numpy array."""
-            spec1 = numpy.vstack((spectrum1.peaks.mz, spectrum1.peaks.intensities)).T
-            spec2 = numpy.vstack((spectrum2.peaks.mz, spectrum2.peaks.intensities)).T
-            assert max(spec1[:, 1]) <= 1, ("Input spectrum1 is not normalized. ",
-                                           "Apply 'normalize_intensities' filter first.")
-            assert max(spec2[:, 1]) <= 1, ("Input spectrum2 is not normalized. ",
-                                           "Apply 'normalize_intensities' filter first.")
-            return spec1, spec2
-
         def get_matching_pairs():
             """Find all pairs of peaks that match within the given tolerance."""
             zero_pairs = collect_peak_pairs(spec1, spec2, self.tolerance, shift=0.0)
@@ -58,22 +79,7 @@ class ModifiedCosine:
             unsorted_matching_pairs = zero_pairs + nonzero_pairs
             return sorted(unsorted_matching_pairs, key=lambda x: x[2], reverse=True)
 
-        def calc_score():
-            """Calculate modified cosine score."""
-            used1 = set()
-            used2 = set()
-            score = 0.0
-            used_matches = []
-            for match in matching_pairs:
-                if not match[0] in used1 and not match[1] in used2:
-                    score += match[2]
[…]
