### 2020-06-29 61c24cb
fix score_best_matches function
