+++ CHANGELOG.md
@@ -11,6 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added similarity score based on comparing parent masses [#79](https://github.com/matchms/matchms/pull/79)
 
+### Changed
+
+- Changed default CosineGreedy function [#86](https://github.com/matchms/matchms/pull/86)
+- Extended readthedocs documentation [#82](https://github.com/matchms/matchms/issues/82)
+
 ## [0.4.0] - 2020-06-11
 
 ### Added
