### 2020-06-30 e7e7983
Update matchms/similarity/ModifiedCosine.py
Co-authored-by: Stefan Verhoeven <s.verhoeven@esciencecenter.nl>
