+++ matchms/similarity/CosineGreedy.py
@@ -1,8 +1,8 @@
 from typing import Tuple
+from matchms.typing import SpectrumType
 from .spectrum_similarity_functions import collect_peak_pairs
 from .spectrum_similarity_functions import get_peaks_array
 from .spectrum_similarity_functions import score_best_matches
-from matchms.typing import SpectrumType
 
 
 class CosineGreedy:
