### 2020-06-30 b875baa
Update CHANGELOG.md
Co-authored-by: Stefan Verhoeven <s.verhoeven@esciencecenter.nl>
