### 2020-06-30 #2398 @lfnothias
Dear editing team of JOSS, I prefer to decline as I am not an expert in software/python. However, let me connect you with the most qualified reviewers for that submission. @bittremieux @kaibioinfo
