### 2020-06-30 #2398 @arfon
Thanks @lfnothias!

@bittremieux @kaibioinfo - would you both be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
