### 2020-06-30 #2398 @arfon
> Yes, I'm willing to review this work.

Thanks @bittremieux! I'll start the main review issue once we have a second reviewer assigned here too. In the meantime, feel free to take a look at our reviewer documentation here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
