### 2020-07-01 bd041fd
Create a new msp file with only 10 spectrums Turned mz and intensity to numpy arrays Added return type to parse_msp_file function Added description for test file Added missing assert test on test file
