### 2020-07-01 2f19b02
Fixing syntaxis issues
