### 2020-07-02 3299817
added functionality to create a spectrum from the usi
