+++ matchms/importing/load_from_usi.py
@@ -1,9 +1,9 @@
-from typing import Optional
 import requests
 import numpy as np
 from matchms import Spectrum
 
-def load_from_usi(usi: str ,server = 'https://metabolomics-usi.ucsd.edu'):
+
+def load_from_usi(usi: str, server = 'https://metabolomics-usi.ucsd.edu'):
     """Load spectrum from metabolomics USI.
 
 USI returns JSON data with keys 'peaks', 'n_peaks' and 'precuror_mz'
@@ -19,7 +19,7 @@ USI returns JSON data with keys 'peaks', 'n_peaks' and 'precuror_mz'
 
     # create the url
     url = server + '/json/?usi=' + usi
-    metadata = {'usi':usi, 'server': server}
+    metadata = {'usi': usi, 'server': server}
     response = requests.get(url)
 
     # extract data and create Spectrum object
@@ -32,9 +32,9 @@ USI returns JSON data with keys 'peaks', 'n_peaks' and 'precuror_mz'
 
         metadata['precursor_mz'] = spectral_data.get('precursor_mz',None)
 
-        s = Spectrum(mz_array,intensity_array,metadata)
+        s = Spectrum(mz_array, intensity_array, metadata)
 
         return s
     except:
-        # failed
+        # failed to unpack json
         return None
