### 2020-07-02 37bc898
and some more style fixes
