+++ matchms/importing/load_from_usi.py
@@ -1,6 +1,6 @@
-import requests
-import numpy as np
 import json
+import numpy as np
+import requests
 from matchms import Spectrum
 
 
