### 2020-07-02 91a099b
added requests to .yml files
