+++ conda/environment-dev.yml
@@ -10,6 +10,7 @@ dependencies:
   - numpy
   - pip
   - pyteomics >=4.2
+  - requests
   - python >=3.7,<3.8
   - pyyaml
   - rdkit >=2020.03.1
