### 2020-07-02 8390ea0
remove CosineGreedyVectorial
