+++ matchms/similarity/__init__.py
@@ -1,6 +1,5 @@
 """similarity module"""
 from .CosineGreedy import CosineGreedy
-from .CosineGreedyVectorial import CosineGreedyVectorial
 from .CosineHungarian import CosineHungarian
 from .FingerprintSimilarityParallel import FingerprintSimilarityParallel
 from .IntersectMz import IntersectMz
@@ -11,7 +10,6 @@ from .ParentmassMatchParallel import ParentmassMatchParallel
 
 __all__ = [
     "CosineGreedy",
-    "CosineGreedyVectorial",
     "CosineHungarian",
     "FingerprintSimilarityParallel",
     "IntersectMz",
