### 2020-07-02 2d09def
Delete CosineGreedyVectorial.py
Replaced by CosineGreedy Numba version
