### 2020-07-02 112a684
use pytest
`python setup.py test` has been deprecated.
