### 2020-07-02 826a691
Erased misplaced documentation from index.rst Added documentation about load_from_msp function in the origin file Added short description for load_from_msp test file
