### 2020-07-02 1bb15f2
added usi tests
