+++ tests/test_usi.py
@@ -1,4 +1,5 @@
-from unittest.mock import Mock, patch
+from unittest.mock import Mock
+from unittest.mock import patch
 import numpy as np
 from matchms import Spectrum
 from matchms.importing.load_from_usi import load_from_usi
