### 2020-07-05 matchms/matchms#94 @kaibioinfo
Regular expression for molecular formulas is ambiguous
The current regular expression for molecular formulas is ambiguous.
`regexp = r"^[C][0-9][0-9A-Z]{4,}$"`
This will match against e.g. Kegg IDs like C01902. Furthermore it won't match any molecular formula that contains lowercases, like C6H4Cl.

I suggest to use
`^(?:[A-Z][a-z]?(?:[1-9][0-9]*)?){2,}$`
instead.
