### 2020-07-05 matchms/matchms#95 @kaibioinfo
Current scope of the software is not clear from the paper
The paper and documentation describes matchms as a general purpose tool for importing mass spectra, cleaning metadata, and calculating similarities.
Such a tool would be, indeed, highly appreciated. There are many databases for small molecule MS, like MassBank, MoNA, and GNPS, and all of them use different data formats. This has historical reasons: the current gold standard format for mass spectrometry is mzML, but this format is very complex and not easy readable in plain text. Most databases developed own formats for storing mass spectra and meta information. MassBank uses its own MSP format, MoNA its own JSON-style format, and GNPS utilizes the MGF format. MGF is a proteomics data format and, as such, not suited for small molecules. Thats why people added metadata tags either as custom fields, or as part of the TITLE field. This is the reason why metadata is not as accessible as in other databases. Furthermore, metadata given by the user is often incomplete or contradictory.

matchms is a tool that can read GNPS MGF files and cleanup the metdata. This seems very similar to what MoNA is doing internally. Furthermore, matchms provides functions for plotting and comparing mass spectra, as well as simple filter functions.

However, the tool only provides the two standard measures for comparing spectra, namely cosine and modified cosine (the latter is known under MANY names).  The implementation of these measures is straight forward, but probably not the most efficient way. Many functionality is missing. There is no weighting of peaks, no removal of information-less peaks. Similarly, the tool only provides very few very simple filter functions, namely: removing low intensities or keeping at max K peaks. Other, more sophisticated, filters are missing, like:
- removing isotopic peaks
- removing peaks which are not decomposable
- fitting noise distribution and removing noise peaks
- removing noise locally via a sliding window
and many more.

I don't want to say that these functions have to be part of the software package. My only concern is that the software package is, based on its functionality, an importer/exporter for GNPS spectra with methods for cleaning up metadata and comparing spectra. You cannot parse spectra from other databases. The JSON importer/exporter, for example, can only read GNPS json, not MoNA json. You cannot parse the gold-standard data format mzML. Most of the "filters" (in my opinion a missleading name) for metadata cleanup are customized for GNPS spectra. For MassBank you don't need to parse adducts from the name and so on, but might have different metadata issues.

I suggest to makes this more clear in the paper. I have seen pending issues for supporting other formats and databases. This is highly appreciated. In its current state, however, the tool is customized for GNPS data and should clearly state this in its documentation and paper.
