### 2020-07-05 matchms/matchms#96 @kaibioinfo
Cosine similarity misses weighting and removal of precursor
* The requirement for normalizing intensities before calculating the cosine is not necessary. The cosine is, by definition, normalized.
* A classical paper about comparing spectra is Stein et al. 1994. There the suggest to
  * take the square root of the intensity, such that the cosine is not driven by the few peaks with highest intensity
  * multiply the intensity by the mass, such that high mass peaks are higher weighted than low mass peaks
* many implementations of cosine also remove peaks around the parent peak (often in a +-20 Da window), such that the parent and uninformative losses like H2O are not dominating the cosine. Instead of deleting them, one could also downweight them.
* peaks behind the precursor mass should not be considered for the cosine. This is possible by using the select_mass filter, but it is strange that nobody enforces this.
