### 2020-07-05 #2411 @kaibioinfo
I added three issues:

One rather major:
Current scope of the software is not clear from the paper (https://github.com/matchms/matchms/issues/95#issue-651066231)

And two minors:
Cosine similarity misses weighting and removal of precursor
https://github.com/matchms/matchms/issues/96#issue-651068504

Regular expression for molecular formulas is ambiguous
https://github.com/matchms/matchms/issues/94#issue-651063454
