### 2020-07-05 matchms/matchms#95 @florian-huber
Many thanks @kaibioinfo for the constructive feedback! That is really very much appreciated and we will try to address all issues you raised.
The next weeks it might go a bit slow due to holiday times, but we certainly planed to include more parsing options anyway (+now having an extra reason). I will also try to address this more clearly in the paper and will let you know once changes have been made.
