### 2020-07-05 e3de63e
add regex suggested by Kai Dührkop
