+++ tests/test_load_from_msp.py
@@ -18,7 +18,7 @@ def test_load_from_msp():
     ])
 
     for k, n in enumerate(spectrum):
-        assert n.get("inchikey") == expected_inchikey[k], "Expected different InChIKey."
+        assert n.get("inchikey").lower() == expected_inchikey[k].lower(), "Expected different InChIKey."
 
 
 test_load_from_msp()
