### 2020-07-08 1c73dfe
fix cosine score denominator
