### 2020-07-08 af77911
updated integration test
