### 2020-07-08 8c156cc
revert to pylint 2.4.4
