+++ conda/environment.yml
@@ -8,6 +8,7 @@ dependencies:
   - matplotlib
   - numba >=0.47
   - numpy
+  - pylint==2.4.4
   - pyteomics >=4.2
   - python >=3.7,<3.8
   - pyyaml
