### 2020-07-08 878afa5
fix sqrt product
