### 2020-07-10 45f8f9a
Downgrade isort, to be compatible with pylint
