### 2020-07-10 a162744
Mock requests directly
