### 2020-07-13 2a60eb1
suggested change to spectrum_similarity_functions.py to enable more general cosine similairites with mz and intensity raised to arbitrary powers
