### 2020-07-13 ccfd910
fixed expected answer in test_modified_cosine_with_mass_shift_5_tolerance_2
