### 2020-07-14 matchms/matchms#100 @bittremieux
pip install
Besides via conda, it would be useful to submit matchms to PyPI as well so that non-conda users can easily install it using pip.
