### 2020-07-15 matchms/matchms#101 @bittremieux
Documentation improvements
The documentation should contain the following information:

- The version(s) of Python that are supported, or at least the minimum version.
- There is no link to the documentation on readthedocs from GitHub.
- [`is not`](https://docs.python.org/3/reference/expressions.html#is) should be used to check whether the reference and query are the same spectrum when iterating over the scores. Instead `==` is used in the readme, on readthedocs, and in the manuscript at the moment.
