### 2020-07-15 #2411 @bittremieux
I would recommend a few more citations to relevant (Python) packages that cover similar functionality, such as [OpenMS](https://doi.org/10.1038/nmeth.3959)/[pyOpenMS](https://doi.org/10.1002/pmic.201300246) and [spectrum_utils](https://doi.org/10.1021/acs.analchem.9b04884) (disclaimer: this is my own package). [MSnbase](https://doi.org/10.1093/bioinformatics/btr645) is a highly relevant R package as well. Given the high overlap between matchms and [spectrum_utils](https://github.com/bittremieux/spectrum_utils/) a more detailed comparison in terms of spectrum processing functionality and computational efficiency might be warranted, although as spectrum_utils developer I leave the final decision about this to @arfon.

Minor comments for the manuscript:

- Abstract: "full quantitative analysis" -> I suggest to rephrase this to avoid confusion with quantitative metabolomics experiments, i.e. where the abundance of the metabolites is measured.
- The correct reference for Numba is: https://doi.org/10.1145/2833157.2833162

I agree with the issues raised by @kaibioinfo, especially on scope. This is not necessarily a major problem, but should be clarified better. Additionally, I have created two more small issues:

- [matchms/matchms#100](https://github.com/matchms/matchms/issues/100): Install matchms using pip.
- [matchms/matchms#101](https://github.com/matchms/matchms/issues/101): Documentation improvements.
