### 2020-07-15 matchms/matchms#100 @sverhoeven
Conda is used mainly because of rdkit usage in the molecule/smile/inchi functions. If matchms was on pypi you would need to install RDKit without using conda. Installing rdkit without conda is hard so we choose to not have a pypi package.

We could make those rdkit dependent functions dynamically check if rdkit is installed and give an error when rdkit is missing, but this could be surprising and hard to document.
