### 2020-07-15 6335709
Removed .lower() from assigned values in Dict when reading msp file Removed unnecessary called to function in test_load_from_msp
