+++ matchms/importing/load_from_msp.py
@@ -33,11 +33,11 @@ def parse_msp_file(filename: str) -> List[dict]:
                     for s in splitted_line[1][2:-1].split('" "'):
                         splitted_line = s.split("=", 1)
                         if splitted_line[0].lower() in params.keys() and splitted_line[0].lower() == 'smiles':
-                            params[splitted_line[0].lower()+"_2"] = splitted_line[1].strip().lower()
+                            params[splitted_line[0].lower()+"_2"] = splitted_line[1].strip()
                         else:
-                            params[splitted_line[0].lower()] = splitted_line[1].strip().lower()
+                            params[splitted_line[0].lower()] = splitted_line[1].strip()
                 else:
-                    params[splitted_line[0].lower()] = splitted_line[1].strip().lower()
+                    params[splitted_line[0].lower()] = splitted_line[1].strip()
             else:
                 # Obtaining the masses and intensities
                 peakscount += 1
