### 2020-07-15 matchms/matchms#100 @bittremieux
I agree, installing RDKit without Conda is very annoying. However, only a minority of people consistently use Conda as package manager (unfortunately).

At least providing the more general way to install matchms using pip gives those users the option to still be able to install it. They'll have to figure out how to get the RDKit dependency themselves.
