### 2020-07-16 36a407e
remove pylint specific version
