+++ conda/meta.yaml
@@ -32,7 +32,6 @@ requirements:
   host:
     - python >=3.7,<3.8
     - pip
-    - pylint==2.4.4
     - pytest-runner
     - setuptools
   run:
