+++ conda/environment-dev.yml
@@ -9,7 +9,6 @@ dependencies:
   - numba >=0.47
   - numpy
   - pip
-    pylint==2.4.4
   - pyteomics >=4.2
   - python >=3.7,<3.8
   - pyyaml
