### 2020-07-16 a0fbc1e
remove pylint specific version
