### 2020-07-16 fcb1a6f
update documentation code examples to fixed cosine score
