### 2020-07-16 e3a9a35
update code examples to fixed cosine
