### 2020-07-16 f375ffe
implement peak powers into score class
