+++ matchms/similarity/ModifiedCosine.py
@@ -46,14 +46,21 @@ class ModifiedCosine:
         Modified cosine score is 0.83 with 1 matched peaks
 
     """
-    def __init__(self, tolerance: float = 0.1):
+    def __init__(self, tolerance: float = 0.1, mz_power: float = 0.0,
+                 intensity_power: float = 1.0):
         """
-        Args:
-        -----
-        tolerance
-            Peaks will be considered a match when <= tolerance a part. Default is 0.1.
+        Parameters
+        ----------
+        tolerance:
+            Peaks will be considered a match when <= tolerance apart. Default is 0.1.
+        mz_power:
+            The power to raise mz to in the cosine function. The default is 0.
+        intensity_power:
+            The power to raise intensity to in the cosine function. The default is 1.
         """
         self.tolerance = tolerance
+        self.mz_power = mz_power
+        self.intensity_power = intensity_power
 
     def __call__(self, spectrum1: SpectrumType, spectrum2: SpectrumType) -> Tuple[float, int]:
         """Calculate modified cosine score between two spectra.
@@ -71,15 +78,20 @@ class ModifiedCosine:
         """
         def get_matching_pairs():
             """Find all pairs of peaks that match within the given tolerance."""
-            zero_pairs = collect_peak_pairs(spec1, spec2, self.tolerance, shift=0.0)
+            zero_pairs = collect_peak_pairs(spec1, spec2, self.tolerance, shift=0.0,
+                                            mz_power=self.mz_power,
+                                            intensity_power=self.intensity_power)
             message = "Precursor_mz missing. Apply 'add_precursor_mz' filter first."
             assert spectrum1.get("precursor_mz") and spectrum2.get("precursor_mz"), message
             mass_shift = spectrum1.get("precursor_mz") - spectrum2.get("precursor_mz")
-            nonzero_pairs = collect_peak_pairs(spec1, spec2, self.tolerance, shift=mass_shift)
+            nonzero_pairs = collect_peak_pairs(spec1, spec2, self.tolerance, shift=mass_shift,
+                                               mz_power=self.mz_power,
+                                               intensity_power=self.intensity_power)
             unsorted_matching_pairs = zero_pairs + nonzero_pairs
             return sorted(unsorted_matching_pairs, key=lambda x: x[2], reverse=True)
 
         spec1 = get_peaks_array(spectrum1)
         spec2 = get_peaks_array(spectrum2)
         matching_pairs = get_matching_pairs()
-        return score_best_matches(matching_pairs, spec1, spec2)
+        return score_best_matches(matching_pairs, spec1, spec2,
+                                  self.mz_power, self.intensity_power)
