### 2020-07-16 696e7d6
linting, changelog, and added test
