+++ CHANGELOG.md
@@ -9,12 +9,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added peak weighting option for cosine score and modified cosine score [#96](https://github.com/matchms/matchms/issues/96)
 - Added similarity score based on comparing parent masses [#79](https://github.com/matchms/matchms/pull/79)
 
 ### Changed
 
 - CosineGreedy function is now numba based [#86](https://github.com/matchms/matchms/pull/86)
 - Extended readthedocs documentation [#82](https://github.com/matchms/matchms/issues/82)
+
+### Fixed
+
+- Fixed incorrect denominator for cosine score normalization [#98](https://github.com/matchms/matchms/pull/98)
 
 ## [0.4.0] - 2020-06-11
 
