### 2020-07-16 bb3c9e2
Update tests/test_usi.py
Co-authored-by: Stefan Verhoeven <s.verhoeven@esciencecenter.nl>
