### 2020-07-16 41db940
Update matchms/importing/load_from_usi.py
Co-authored-by: Stefan Verhoeven <s.verhoeven@esciencecenter.nl>
