### 2020-07-27 53572bd
Update CHANGELOG.md
Co-authored-by: Stefan Verhoeven <s.verhoeven@esciencecenter.nl>
