### 2020-07-27 72b68e4
Update CHANGELOG.md
Co-authored-by: Stefan Verhoeven <s.verhoeven@esciencecenter.nl>
