### 2020-07-27 70fd179
Update matchms/similarity/spectrum_similarity_functions.py
Co-authored-by: Stefan Verhoeven <s.verhoeven@esciencecenter.nl>
