### 2020-07-27 137f51d
remove normalization check
