### 2020-07-27 matchms/matchms#94 @florian-huber
Many thanks @kaibioinfo ! We added the improved regex that you suggested --> #97
