### 2020-07-27 7a83b7f
linting changelog and rename test_usi to test_load_from_usi
