### 2020-07-27 cfc4c69
add code example and fix bug
