+++ matchms/importing/load_from_usi.py
@@ -4,14 +4,21 @@ import requests
 from ..Spectrum import Spectrum
 
 
-def load_from_usi(usi: str, server='https://metabolomics-usi.ucsd.edu'):
+def load_from_usi(usi: str, server: str = "https://metabolomics-usi.ucsd.edu"):
     """Load spectrum from metabolomics USI.
 
-USI returns JSON data with keys 'peaks', 'n_peaks' and 'precuror_mz'
+    USI returns JSON data with keys "peaks", "n_peaks" and "precuror_mz"
+
+    .. code-block:: python
+
+        from matchms.importing import load_from_usi
+
+        spectrum = load_from_usi("mzspec:MASSBANK::accession:SM858102")
+        print(f"Found spectrum with precursor m/z of {spectrum.get("precursor_mz"):.2f}.")
 
     Args:
     ----
-    usi: str
+    usi:
         Provide the usi.
 
     server: string
@@ -19,25 +26,25 @@ USI returns JSON data with keys 'peaks', 'n_peaks' and 'precuror_mz'
     """
 
     # Create the url
-    url = server + '/json/?usi=' + usi
-    metadata = {'usi': usi, 'server': server}
+    url = server + "/json/?usi=" + usi
+    metadata = {"usi": usi, "server": server}
     response = requests.get(url)
 
-    if response.statuscode == 404:
+    if response.status_code == 404:
         return None
     # Extract data and create Spectrum object
     try:
         spectral_data = response.json()
-        if spectral_data is None or 'peaks' not in spectral_data:
+        if spectral_data is None or "peaks" not in spectral_data:
             return None
-        peaks = spectral_data['peaks']
+        peaks = spectral_data["peaks"]
         if len(peaks) == 0:
             return None
         mz_list, intensity_list = zip(*peaks)
         mz_array = np.array(mz_list)
         intensity_array = np.array(intensity_list)
 
-        metadata['precursor_mz'] = spectral_data.get('precursor_mz', None)
+        metadata["precursor_mz"] = spectral_data.get("precursor_mz", None)
 
         s = Spectrum(mz_array, intensity_array, metadata)
 
