### 2020-07-27 391f07a
paper update to address reviewer comments
