+++ paper/paper.bib
@@ -141,16 +141,24 @@
 }
 
 
-@InProceedings{LLVM:CGO04,
-     Author  = {Chris Lattner and Vikram Adve},
-     Title = {{LLVM}: A Compilation Framework for Lifelong Program
-Analysis and Transformation},
-     Booktitle = CGO,
-     Address = {San Jose, CA, USA},
-     Month = {Mar},
-     Year  = {2004},
-     pages       = {75--88},
+@inproceedings{lam_numba_2015,
+	address = {Austin, Texas},
+	series = {{LLVM} '15},
+	title = {Numba: a {LLVM}-based {Python} {JIT} compiler},
+	isbn = {978-1-4503-4005-2},
+	shorttitle = {Numba},
+	url = {https://doi.org/10.1145/2833157.2833162},
+	doi = {10.1145/2833157.2833162},
+	urldate = {2020-07-27},
+	booktitle = {Proceedings of the {Second} {Workshop} on the {LLVM} {Compiler} {Infrastructure} in {HPC}},
+	publisher = {Association for Computing Machinery},
+	author = {Lam, Siu Kwan and Pitrou, Antoine and Seibert, Stanley},
+	month = nov,
+	year = {2015},
+	keywords = {compiler, LLVM, Python},
+	pages = {1--6}
 }
+
 
 
 @article{watrous_mass_2012,
@@ -296,4 +304,63 @@ note = {[Online; accessed 25-June-2020]},
 	keywords = {distributed database, MassBank, metabolite, public database, spectral similarity},
 	pages = {703--714},
 	file = {Snapshot:C\:\\Users\\FlorianHuber\\Zotero\\storage\\ZJ9TNXKY\\jms.html:text/html}
+}
+
+
+@article{rost_pyopenms_2014,
+	title = {{pyOpenMS}: {A} {Python}-based interface to the {OpenMS} mass-spectrometry algorithm library},
+	volume = {14},
+	issn = {1615-9861},
+	shorttitle = {{pyOpenMS}},
+	url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/pmic.201300246},
+	doi = {10.1002/pmic.201300246},
+	language = {en},
+	number = {1},
+	urldate = {2020-07-27},
+	journal = {PROTEOMICS},
+	author = {RÃ¶st, Hannes L. and Schmitt, Uwe and Aebersold, Ruedi and MalmstrÃ¶m, Lars},
+	year = {2014},
+	note = {\_eprint: https://onlinelibrary.wiley.com/doi/pdf/10.1002/pmic.201300246},
+	keywords = {Bioinformatics, Data mining, Mass spectra, Proteome maps, Python, Software},
+	pages = {74--77},
+	file = {Snapshot:C\:\\Users\\FlorianHuber\\Zotero\\storage\\3M4UVEL4\\pmic.html:text/html}
+}
+
+@article{rost_openms_2016,
+	title = {{OpenMS}: a flexible open-source software platform for mass spectrometry data analysis},
+	volume = {13},
+	copyright = {2016 Nature Publishing Group, a division of Macmillan Publishers Limited. All Rights Reserved.},
+	issn = {1548-7105},
+	shorttitle = {{OpenMS}},
+	url = {https://www.nature.com/articles/nmeth.3959},
+	doi = {10.1038/nmeth.3959},
+	language = {en},
+	number = {9},
+	urldate = {2020-07-27},
+	journal = {Nature Methods},
[…]
