+++ paper/paper.bib
@@ -318,7 +318,7 @@ note = {[Online; accessed 25-June-2020]},
 	number = {1},
 	urldate = {2020-07-27},
 	journal = {PROTEOMICS},
-	author = {RÃ¶st, Hannes L. and Schmitt, Uwe and Aebersold, Ruedi and MalmstrÃ¶m, Lars},
+	author = {R{\"o}st, Hannes L. and Schmitt, Uwe and Aebersold, Ruedi and Malmstr{\"o}m, Lars},
 	year = {2014},
 	note = {\_eprint: https://onlinelibrary.wiley.com/doi/pdf/10.1002/pmic.201300246},
 	keywords = {Bioinformatics, Data mining, Mass spectra, Proteome maps, Python, Software},
@@ -338,7 +338,7 @@ note = {[Online; accessed 25-June-2020]},
 	number = {9},
 	urldate = {2020-07-27},
 	journal = {Nature Methods},
-	author = {RÃ¶st, Hannes L. and Sachsenberg, Timo and Aiche, Stephan and Bielow, Chris and Weisser, Hendrik and Aicheler, Fabian and Andreotti, Sandro and Ehrlich, Hans-Christian and Gutenbrunner, Petra and Kenar, Erhan and Liang, Xiao and Nahnsen, Sven and Nilse, Lars and Pfeuffer, Julianus and Rosenberger, George and Rurik, Marc and Schmitt, Uwe and Veit, Johannes and Walzer, Mathias and Wojnar, David and Wolski, Witold E. and Schilling, Oliver and Choudhary, Jyoti S. and MalmstrÃ¶m, Lars and Aebersold, Ruedi and Reinert, Knut and Kohlbacher, Oliver},
+	author = {R{\"o}st, Hannes L. and Sachsenberg, Timo and Aiche, Stephan and Bielow, Chris and Weisser, Hendrik and Aicheler, Fabian and Andreotti, Sandro and Ehrlich, Hans-Christian and Gutenbrunner, Petra and Kenar, Erhan and Liang, Xiao and Nahnsen, Sven and Nilse, Lars and Pfeuffer, Julianus and Rosenberger, George and Rurik, Marc and Schmitt, Uwe and Veit, Johannes and Walzer, Mathias and Wojnar, David and Wolski, Witold E. and Schilling, Oliver and Choudhary, Jyoti S. and Malmstr{\"o}m, Lars and Aebersold, Ruedi and Reinert, Knut and Kohlbacher, Oliver},
 	month = sep,
 	year = {2016},
 	note = {Number: 9
