### 2020-07-27 0616e91
Update matchms/importing/load_from_msp.py
Thank you for the suggestion, I totally ignored that.

Co-authored-by: Stefan Verhoeven <s.verhoeven@esciencecenter.nl>
