### 2020-07-27 dc2043b
Update matchms/importing/load_from_msp.py
Co-authored-by: Stefan Verhoeven <s.verhoeven@esciencecenter.nl>
