### 2020-07-28 a02c8dc
minor linting
