### 2020-07-28 matchms/matchms#96 @florian-huber
Working on it...
So far:
- removed the non-essential normalization requirement (PR #103).
- added option to Cosine score for raising peaks by power of m/z and/or intensity (PR #99 )
