### 2020-07-29 f990995
Added additional filtering methods
