+++ paper/paper.bib
@@ -364,3 +364,23 @@ Publisher: Nature Publishing Group},
 	pages = {659--661},
 	file = {Accepted Version:C\:\\Users\\FlorianHuber\\Zotero\\storage\\262VVEED\\Bittremieux - 2020 - spectrum_utils A Python Package for Mass Spectrom.pdf:application/pdf}
 }
+
+@article{kosters_pymzml_2018,
+	title = {{pymzML} v2.0: introducing a highly compressed and seekable gzip format},
+	volume = {34},
+	issn = {1367-4803},
+	shorttitle = {{pymzML} v2.0},
+	url = {https://academic.oup.com/bioinformatics/article/34/14/2513/4831092},
+	doi = {10.1093/bioinformatics/bty046},
+	abstract = {AbstractMotivation.  In the new release of pymzML (v2.0), we have optimized the speed of this established tool for mass spectrometry data analysis to adapt to i},
+	language = {en},
+	number = {14},
+	urldate = {2020-07-30},
+	journal = {Bioinformatics},
+	author = {KÃ¶sters, M. and Leufken, J. and Schulze, S. and Sugimoto, K. and Klein, J. and Zahedi, R. P. and Hippler, M. and Leidel, S. A. and Fufezan, C.},
+	month = jul,
+	year = {2018},
+	note = {Publisher: Oxford Academic},
+	pages = {2513--2514},
+	file = {Full Text PDF:C\:\\Users\\FlorianHuber\\Zotero\\storage\\MFI5YTGN\\KÃ¶sters et al. - 2018 - pymzML v2.0 introducing a highly compressed and s.pdf:application/pdf;Snapshot:C\:\\Users\\FlorianHuber\\Zotero\\storage\\E9US9R68\\4831092.html:text/html}
+}
