+++ paper/paper.bib
@@ -176,7 +176,6 @@
 	pmcid = {PMC3387089},
 	keywords = {Bacillus subtilis, Base Sequence, Mass Spectrometry, Multigene Family, Mycobacterium smegmatis, Pseudomonas aeruginosa, Streptomyces coelicolor},
 	pages = {E1743--1752},
-	file = {Full Text:C\:\\Users\\FlorianHuber\\Zotero\\storage\\FLBCPG88\\Watrous et al. - 2012 - Mass spectral molecular networking of living micro.pdf:application/pdf}
 }
 
 
@@ -219,7 +218,6 @@
 	note = {Conference Name: Computing in Science Engineering},
 	keywords = {Arrays, Computational efficiency, data structures, Finite element methods, high level language, high level languages, mathematics computing, numerical analysis, Numerical analysis, numerical computation, numerical computations, numerical data, NumPy, numpy array, Performance evaluation, programming libraries, Python, Python programming language, Resource management, scientific programming, Vector quantization},
 	pages = {22--30},
-	file = {Full Text:C\:\\Users\\FlorianHuber\\Zotero\\storage\\GFNMR5FK\\van der Walt et al. - 2011 - The NumPy Array A Structure for Efficient Numeric.pdf:application/pdf;IEEE Xplore Abstract Record:C\:\\Users\\FlorianHuber\\Zotero\\storage\\WQ6GGPUU\\5725236.html:text/html}
 }
 
 
@@ -283,7 +281,6 @@ note = {[Online; accessed 25-June-2020]},
 	year = {2020},
 	note = {Publisher: Oxford Academic},
 	pages = {D440--D444},
-	file = {Full Text PDF:C\:\\Users\\FlorianHuber\\Zotero\\storage\\ME2S77ML\\Haug et al. - 2020 - MetaboLights a resource evolving in response to t.pdf:application/pdf}
 }
 
 
@@ -303,7 +300,6 @@ note = {[Online; accessed 25-June-2020]},
 	year = {2010},
 	keywords = {distributed database, MassBank, metabolite, public database, spectral similarity},
 	pages = {703--714},
-	file = {Snapshot:C\:\\Users\\FlorianHuber\\Zotero\\storage\\ZJ9TNXKY\\jms.html:text/html}
 }
 
 
@@ -323,7 +319,6 @@ note = {[Online; accessed 25-June-2020]},
 	note = {\_eprint: https://onlinelibrary.wiley.com/doi/pdf/10.1002/pmic.201300246},
 	keywords = {Bioinformatics, Data mining, Mass spectra, Proteome maps, Python, Software},
 	pages = {74--77},
-	file = {Snapshot:C\:\\Users\\FlorianHuber\\Zotero\\storage\\3M4UVEL4\\pmic.html:text/html}
 }
 
 @article{rost_openms_2016,
@@ -344,7 +339,6 @@ note = {[Online; accessed 25-June-2020]},
 	note = {Number: 9
 Publisher: Nature Publishing Group},
 	pages = {741--748},
-	file = {Accepted Version:C\:\\Users\\FlorianHuber\\Zotero\\storage\\VZBRLNH8\\RÃ¶st et al. - 2016 - OpenMS a flexible open-source software platform f.pdf:application/pdf;Snapshot:C\:\\Users\\FlorianHuber\\Zotero\\storage\\UDQ45KZA\\nmeth.html:text/html}
 }
 
 @article{bittremieux_spectrum_utils_2020,
@@ -362,7 +356,6 @@ Publisher: Nature Publishing Group},
 	year = {2020},
[…]
