### 2020-07-31 002beb0
Add link to readthedocs and required Python version
