+++ paper/paper.md
@@ -113,7 +113,7 @@ for score in scores:
     (reference, query, score, n_matching) = score
     # Ignore scores between same spectrum and
     # pairs which have less than 20 peaks in common
-    if reference != query and n_matching >= 20:
+    if reference is not query and n_matching >= 20:
         print(f"Reference scan id: {reference.metadata['scans']}")
         print(f"Query scan id: {query.metadata['scans']}")
         print(f"Score: {score:.4f}")
