### 2020-07-31 matchms/matchms#101 @florian-huber
Many thanks for the good hints!
- First two items were addressed in PR #108 .
- And ``!=`` was changed to ``is not`` in code examples (PR #109 and ongoing PR #105 ).
