### 2020-08-03 2ca3432
linting and moving helper functions to utils
