### 2020-08-03 db8de98
remove redundant code
