+++ matchms/importing/__init__.py
@@ -3,6 +3,7 @@ from .load_from_json import load_from_json
 from .load_from_mgf import load_from_mgf
 from .load_from_msp import load_from_msp
 from .load_from_mzml import load_from_mzml
+from .load_from_mzxml import load_from_mzxml
 from .load_from_usi import load_from_usi
 
 
@@ -11,6 +12,7 @@ __all__ = [
     "load_from_mgf",
     "load_from_msp",
     "load_from_mzml",
+    "load_from_mzxml",
     "load_from_usi",
     "load_adducts"
 ]
