### 2020-08-03 8778f01
try linting 2-line if statement...
