+++ matchms/importing/load_from_mzxml.py
@@ -29,8 +29,8 @@ def load_from_mzxml(filename: str, ms_level: int = 2) -> Generator[Spectrum, Non
         Specify which ms level to import. Default is 2.
     """
     for pyteomics_spectrum in mzxml.read(filename, dtype=dict):
-        if "ms level" in pyteomics_spectrum and pyteomics_spectrum["ms level"] == ms_level \
-        or "msLevel" in pyteomics_spectrum and pyteomics_spectrum["msLevel"] == ms_level:
+        if ("ms level" in pyteomics_spectrum and pyteomics_spectrum["ms level"] == ms_level
+                or "msLevel" in pyteomics_spectrum and pyteomics_spectrum["msLevel"] == ms_level):
             metadata = parse_mzml_mzxml_metadata(pyteomics_spectrum)
             mz = numpy.asarray(pyteomics_spectrum["m/z array"], dtype="float")
             intensities = numpy.asarray(pyteomics_spectrum["intensity array"], dtype="float")
