### 2020-08-03 ae47060
try fixing test failing under ubuntu
