+++ tests/test_load_from_mzxml.py
@@ -7,7 +7,7 @@ def test_load_from_mzxml():
 
     module_root = os.path.dirname(__file__)
     mzxml_file = os.path.join(module_root, "testdata.mzxml")
-    spectrum = next(load_from_mzxml(mzxml_file))
+    spectrum = list(load_from_mzxml(mzxml_file))[0]
 
     assert int(spectrum.get("precursor_mz")) == 343, "Expected different precursor m/z"
     assert spectrum.get("charge") == -1, "Expected different charge."
