### 2020-08-03 c4c4137
add better check for empty peak arrays
