### 2020-08-03 91cb7d6
simplify helper function
