### 2020-08-03 c744306
Update docstring of parsing function.
