+++ paper/paper.bib
@@ -376,3 +376,19 @@ Publisher: Nature Publishing Group},
 	note = {Publisher: Oxford Academic},
 	pages = {2513--2514},
 }
+
+@article{wang_interactive_2020,
+	title = {Interactive {MS}/{MS} {Visualization} with the {Metabolomics} {Spectrum} {Resolver} {Web} {Service}},
+	copyright = {Â© 2020, Posted by Cold Spring Harbor Laboratory. This pre-print is available under a Creative Commons License (Attribution-NonCommercial 4.0 International), CC BY-NC 4.0, as described at http://creativecommons.org/licenses/by-nc/4.0/},
+	url = {https://www.biorxiv.org/content/10.1101/2020.05.09.086066v1},
+	doi = {10.1101/2020.05.09.086066},
+	language = {en},
+	urldate = {2020-08-03},
+	journal = {bioRxiv},
+	author = {Wang, Mingxun and Rogers, Simon and Bittremieux, Wout and Chen, Christopher and Dorrestein, Pieter C. and Schymanski, Emma L. and Schulze, Tobias and Neumann, Steffen and Meier, Rene},
+	month = may,
+	year = {2020},
+	note = {Publisher: Cold Spring Harbor Laboratory
+Section: New Results},
+	pages = {2020.05.09.086066},
+}
