### 2020-08-03 matchms/matchms#100 @florian-huber
We can indeed consider to add a "matchms light" version to pip.
Functions that would be affected seem to be:
### from filtering
- ``add_fingerprint``
- ``derive_inchikey_from_inchi``
- ``derive_inchikey_from_smiles``
- ``derive_smiles_from_inchi``
- ``repair_inchi_inchikey_smiles``
### from similarity
- ``FingerprintSimilarity``

However, this seems to require an own sprint. In particular also to properly do the packaging and CI.
For the matchms manuscript that won't be do-able I fear.
