### 2020-08-03 102a136
add intensity and mz power options to Cosine Hungarian
