### 2020-08-03 270f4ba
Remove docstring comment.
Co-authored-by: Stefan Verhoeven <s.verhoeven@esciencecenter.nl>
