### 2020-08-03 dc76166
Clean up docstring bullet points
Co-authored-by: Stefan Verhoeven <s.verhoeven@esciencecenter.nl>
