### 2020-08-04 32e2c25
add scan number
Co-authored-by: Stefan Verhoeven <s.verhoeven@esciencecenter.nl>
