### 2020-08-04 8cb8b8b
check if num in dictionary
