### 2020-08-04 ff0e3c3
Update tests/test_load_from_mzxml.py
Co-authored-by: Stefan Verhoeven <s.verhoeven@esciencecenter.nl>
