### 2020-08-04 1a37dcb
add test using peak powers
