+++ CHANGELOG.md
@@ -10,7 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Read msp files to create Spectrum objects from it [#102](https://github.com/matchms/matchms/pull/102)
-- Peak weighting option for cosine score and modified cosine score [#96](https://github.com/matchms/matchms/issues/96)
+- Peak weighting option for CosineGreedy and ModifiedCosine score [#96](https://github.com/matchms/matchms/issues/96)
+- Peak weighting option for CosineHungarian score [#112](https://github.com/matchms/matchms/pull/112)
 - Similarity score based on comparing parent masses [#79](https://github.com/matchms/matchms/pull/79)
 - Method for instantiating a spectrum from the metabolomics USI [#93](https://github.com/matchms/matchms/pull/93)
 
