### 2020-08-04 45592ec
fix normalization
