### 2020-08-04 28cfa68
update tests to correct denominator
