+++ conda/meta.yaml
@@ -1,5 +1,5 @@
 {% set name = "matchms" %}
-{% set version = "0.4.0" %}
+{% set version = "0.5.0" %}
 
 package:
   name: {{ name|lower }}
