+++ setup.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 0.4.0
+current_version = 0.5.0
 
 [bumpversion:file:conda/meta.yaml]
 search = set version = "{current_version}"
