### 2020-08-05 5e5a734
Add installation instruction to readthedocs
