### 2020-08-05 #2411 @florian-huber
Thanks @arfon and in particular @bittremieux and @kaibioinfo for reviewing.
The raised comments were very valuable and helped us improve the package.

We generally agree with the feedback we received, and think that we could address most issues.
The main exception is making matchms a pip installable package (https://github.com/matchms/matchms/issues/100), which is complicated because matchms relies on ``rdkit``, a conda package. Still, we hope to address this in the future, maybe releasing a lightweight matchms version for pip.

Other than that, we incorporated https://github.com/matchms/matchms/issues/94, https://github.com/matchms/matchms/issues/101 and the minor comments by @bittremieux above. I believe we addressed the main points of https://github.com/matchms/matchms/issues/96 by fixing and extending the cosine score functions. In addition, we currently work on providing more filter functions do address the remaining points (PR https://github.com/matchms/matchms/pull/115). Since most of the still missing functionality could -though cumbersome- be achieve using present filters, this might be less urgent with respect to the JOSS manuscript.

Finally, the major point raised by @kaibioinfo in https://github.com/matchms/matchms/issues/95 (overlapping with recommendation by @bittremieux  to make the scope clearer) was addressed in two ways.
We extended the importing options of matchms, now including mzML, mzXML, msp, and metabolomics-USI. In addition, we now refer to more related packages and state more clearly the limitations of matchms (see https://github.com/matchms/matchms/issues/95 for more detailed description).

### Manuscript and code update:
- the mentioned changes were made into a new matchms release (version 0.5.0).
- the edited paper and bibliography was added to the master branch.
- we added one additional author (Efraín Manuel Villanueva Castilla, who contributed code to this latest release).

Please let us know if that is sufficiently addressing the feedback, or further changes need to be made.
Many thanks!
