### 2020-08-08 b3a426a
Fixed linting issues
