### 2020-08-08 3a98040
Fixed more linting issues
