+++ matchms/filtering/remove_peaks_within_window.py
@@ -1,5 +1,6 @@
 from ..Spikes import Spikes
 from ..typing import SpectrumType
+import numpy
 
 
 def remove_peaks_within_window(spectrum: SpectrumType, k: int = 6,
