### 2020-08-10 dd0bced
Update matchms/filtering/remove_spectra_below.py
Co-authored-by: Florian Huber <36473328+florian-huber@users.noreply.github.com>
