+++ matchms/filtering/remove_spectra_below.py
@@ -1,7 +1,7 @@
 from ..typing import SpectrumType
 
 
-def remove_spectra_below(spectrum: SpectrumType, max_mz: float = 1000) -> SpectrumType:
+def remove_spectra_above_mz(spectrum: SpectrumType, max_mz: float = 1000) -> SpectrumType:
 
     """Returns none if the precursor_mz of a spectrum is above
        max_mz.
