### 2020-08-10 f9b12f6
Update matchms/filtering/remove_spectra_within_tolerance.py
Co-authored-by: Florian Huber <36473328+florian-huber@users.noreply.github.com>
