### 2020-08-10 890f9dd
Update matchms/filtering/remove_spectra_above_intensity.py
Co-authored-by: Florian Huber <36473328+florian-huber@users.noreply.github.com>
