### 2020-08-10 52792c0
Update remove_spectra_below_intensity.py
