+++ matchms/filtering/remove_spectra_few_high_peaks.py
@@ -3,7 +3,7 @@ from . import select_by_relative_intensity
 
 
 def remove_spectra_few_high_peaks(spectrum: SpectrumType, no_peaks: int = 5,
-                                   intensity_percent: float = 2) -> SpectrumType:
+                                  intensity_percent: float = 2) -> SpectrumType:
 
     """Returns none if the number of peaks with relative intensity
        above intensity_percent is less than no_peaks.
