### 2020-08-10 e2f7662
Updated docstring
