### 2020-08-10 54231df
Updated docstring
