### 2020-08-10 c8dce63
Updated docstring
