### 2020-08-11 6c21e79
Limit sphinx version to <3.2.0
