+++ conda/environment-dev.yml
@@ -25,7 +25,7 @@ dependencies:
     - pytest-runner
     - recommonmark
     - ruamel.yaml
-    - sphinx
+    - sphinx>=3.0.0<3.2.0
     - sphinx_rtd_theme
     - sphinxcontrib-apidoc
     - yapf
