### 2020-08-11 a5f969d
add archspec for testing
