+++ conda/environment-dev.yml
@@ -4,7 +4,6 @@ channels:
   - conda-forge
   - defaults
 dependencies:
-  - archspec
   - gensim >=3.8.0
   - matplotlib
   - numba >=0.47
@@ -26,7 +25,7 @@ dependencies:
     - pytest-runner
     - recommonmark
     - ruamel.yaml
-    - sphinx==3.1.2
+    - sphinx>=3.0.0,<3.2.0
     - sphinx_rtd_theme
     - sphinxcontrib-apidoc
     - yapf
