### 2020-08-12 8a8234f
add docstrings to filter functions
