+++ matchms/filtering/select_by_mz.py
@@ -3,8 +3,17 @@ from ..Spikes import Spikes
 from ..typing import SpectrumType
 
 
-def select_by_mz(spectrum_in: SpectrumType, mz_from=0.0, mz_to=1000.0) -> SpectrumType:
+def select_by_mz(spectrum_in: SpectrumType, mz_from: float = 0.0,
+                 mz_to: float = 1000.0) -> SpectrumType:
+    """Keep only peaks between mz_from and mz_to (keep if mz_from >= m/z >= mz_to).
 
+    Args:
+    -----
+    mz_from:
+        Set lower threshold for m/z peak positions. Default is 0.0.
+    mz_to:
+        Set upper threshold for m/z peak positions. Default is 1000.0.
+    """
     if spectrum_in is None:
         return None
 
