### 2020-08-12 e113e07
add further missing docstrings
