+++ matchms/filtering/harmonize_undefined_inchi.py
@@ -1,8 +1,19 @@
+from typing import List
 from ..typing import SpectrumType
 
 
-def harmonize_undefined_inchi(spectrum_in: SpectrumType, undefined="", aliases=None):
-
+def harmonize_undefined_inchi(spectrum_in: SpectrumType, undefined: str = "",
+                              aliases: List[str] = None) -> SpectrumType:
+    """Replace all aliases for empty/undefined inchi entries by ``undefined``.
+    
+    Args:
+    -----
+    undefined:
+        Give desired entry for undefined inchi fields. Default is "".
+    aliases:
+        Enter list of strings that are expected to represent undefined entries.
+        Default is ["", "N/A", "NA", "n/a"].
+    """
     if spectrum_in is None:
         return None
 
