+++ matchms/filtering/derive_ionmode.py
@@ -2,7 +2,7 @@ from ..importing import load_adducts
 from ..typing import SpectrumType
 
 
-def derive_ionmode(spectrum_in: SpectrumType, adducts_filename=None) -> SpectrumType:
+def derive_ionmode(spectrum_in: SpectrumType, adducts_filename: str = None) -> SpectrumType:
     """Derive missing ionmode based on adduct.
 
     MGF files do not always provide a correct ionmode. This function reads
@@ -13,6 +13,8 @@ def derive_ionmode(spectrum_in: SpectrumType, adducts_filename=None) -> Spectrum
     -----
     spectrum:
         Input spectrum.
+    adducts_filename:
+        Load known adducts from file, if filename is given. Default is None.
 
     Returns:
     --------
