+++ matchms/filtering/harmonize_undefined_inchi.py
@@ -5,7 +5,7 @@ from ..typing import SpectrumType
 def harmonize_undefined_inchi(spectrum_in: SpectrumType, undefined: str = "",
                               aliases: List[str] = None) -> SpectrumType:
     """Replace all aliases for empty/undefined inchi entries by ``undefined``.
-    
+
     Args:
     -----
     undefined:
