+++ matchms/filtering/harmonize_undefined_smiles.py
@@ -2,10 +2,10 @@ from typing import List
 from ..typing import SpectrumType
 
 
-def harmonize_undefined_smiles(spectrum_in: SpectrumType, undefined="",
-                              aliases: List[str] = None) -> SpectrumType:
+def harmonize_undefined_smiles(spectrum_in: SpectrumType, undefined: str = "",
+                               aliases: List[str] = None) -> SpectrumType:
     """Replace all aliases for empty/undefined smiles entries by ``undefined``.
-    
+
     Args:
     -----
     undefined:
