### 2020-08-12 fecf0f5
add new filters to __init__
