+++ matchms/filtering/__init__.py
@@ -1,6 +1,10 @@
-"""Function to process mass spectra.
+"""Functions for processing mass spectra.
+Provided functions will usually only perform a single action to a spectrum.
+This can be changes or corrections of metadata, or peak filtering.
+More complicated processing pipelines can be build by stacking several of
+the provided filters.
 
-Example of how to use the provided filter functions:
+Example of how to use a single filter function:
 
 .. code-block:: python
 
