### 2020-08-13 1bb3d57
Update matchms/filtering/derive_adduct_from_name.py
Co-authored-by: Stefan Verhoeven <s.verhoeven@esciencecenter.nl>
