### 2020-08-13 40bb6d5
Update matchms/filtering/harmonize_undefined_inchi.py
Co-authored-by: Stefan Verhoeven <s.verhoeven@esciencecenter.nl>
