+++ matchms/filtering/harmonize_undefined_inchi.py
@@ -4,7 +4,7 @@ from ..typing import SpectrumType
 
 def harmonize_undefined_inchi(spectrum_in: SpectrumType, undefined: str = "",
                               aliases: List[str] = None) -> SpectrumType:
-    """Replace all aliases for empty/undefined inchi entries by ``undefined``.
+    """Replace all aliases for empty/undefined inchi entries by value of ``undefined`` argument.
 
     Args:
     -----
