### 2020-08-13 140ac2f
Update matchms/filtering/select_by_intensity.py
Co-authored-by: Stefan Verhoeven <s.verhoeven@esciencecenter.nl>
