+++ matchms/filtering/select_by_intensity.py
@@ -7,7 +7,7 @@ def select_by_intensity(spectrum_in: SpectrumType, intensity_from: float = 10.0,
                         intensity_to: float = 200.0) -> SpectrumType:
     """Keep only peaks within set intensity range (keep if
     intensity_from >= intensity >= intensity_to). In most cases it is adviced to
-    use ``select_by_relative_intensity`` function instead.
+    use :py:func:`select_by_relative_intensity` function instead.
 
     Args:
     -----
