+++ matchms/filtering/set_ionmode_na_when_missing.py
@@ -2,7 +2,7 @@ from ..typing import SpectrumType
 
 
 def set_ionmode_na_when_missing(spectrum_in: SpectrumType) -> SpectrumType:
-
+    """Create "ionmode" entry of "n/a" if field does not yet exist in metadata."""
     if spectrum_in is None:
         return None
 
