### 2020-08-13 aaea72a
make code example testable
