### 2020-08-13 24850e3
mention correct filters in docstring
