+++ matchms/filtering/__init__.py
@@ -13,11 +13,11 @@ Example of how to use a single filter function:
     from matchms.filtering import normalize_intensities
 
     spectrum = Spectrum(mz=np.array([100, 120, 150, 200.]),
-                          intensities=np.array([200.0, 300.0, 50.0, 1.0]),
-                          metadata={'id': 'spectrum1'})
+                        intensities=np.array([200.0, 300.0, 50.0, 1.0]),
+                        metadata={'id': 'spectrum1'})
     spectrum_filtered = normalize_intensities(spectrum)
 
-    max_intensity = spectrum_filtered.peaks.intensities.max())
+    max_intensity = spectrum_filtered.peaks.intensities.max()
     print(f"Maximum intensity is {max_intensity:.2f}")
 
     Should output
