### 2020-08-13 matchms/matchms#101 @florian-huber
This has been fixed, see above comment.
