### 2020-08-13 bf14f29
couple of typos; style edits
