+++ matchms/importing/load_from_msp.py
@@ -63,11 +63,14 @@ def load_from_msp(filename: str) -> Generator[Spectrum, None, None]:
     Function that reads a .msp file and converts the info
     in :py:class:`~matchms.Spectrum.Spectrum` objects.
 
-    Args:
-        filename: path of the msp file
+    Parameters
+    ----------
+    filename:
+        Path of the msp file.
 
-    Yields:
-        Yield a spectrum object with the data of the msp file
+    Returns
+    -------
+    Yield a spectrum object with the data of the msp file
 
 
     Example:
