### 2020-08-14 110fc8f
Switched some more docstrings to numpy style
