### 2020-08-14 f2dce37
Added tests for remove_peaks_around_precursor_mz
