### 2020-08-15 e95914d
Use is to check for None spectrum
