+++ tests/test_require_precursor_below_mz.py
@@ -35,4 +35,4 @@ def test_require_precursor_below_mz_max_50():
 
     spectrum = require_precursor_below_mz(spectrum_in, max_mz = 50)
 
-    assert spectrum == None, "Expected spectrum to be None."
+    assert spectrum is None, "Expected spectrum to be None."
