### 2020-08-17 8e7f1b4
add submodule docstrings
