+++ matchms/exporting/__init__.py
@@ -1,3 +1,8 @@
+"""Functions for exporting mass spectral data.
+
+:meth:`~matchms.Spectrum` (individual ones, or lists) can be exported to
+json or mgf files.
+"""
 from .save_as_json import save_as_json
 from .save_as_mgf import save_as_mgf
 
