### 2020-08-17 6353e4d
try adding figure to filtering
