+++ matchms/importing/__init__.py
@@ -1,12 +1,14 @@
-"""Functions for importing mass spectral data.
+"""
+Functions for importing mass spectral data
+##########################################
 
 Matchms provides a import functions for several commonly used data types, such
-as _.mzML_, _.mzXML_, _.mgf_, or _.msp_. It is also possible to load data from
-_.json_ files (tested for json files from GNPS or json files made with matchms).
+as *.mzML*, *.mzXML*, *.mgf*, or *.msp*. It is also possible to load data from
+*.json* files (tested for json files from GNPS or json files made with matchms).
 Another option is to load spectra based on a unique identifier (USI)
-(:meth:`~matchms.simporting.load_from_usi`). For more extensive import options
-we recommend building custom importers using __pyteomics__
-(https://github.com/levitsky/pyteomics) or __pymzml__ (https://github.com/pymzml/pymzML).
+(:meth:`~matchms.importing.load_from_usi`). For more extensive import options
+we recommend building custom importers using **pyteomics**
+(https://github.com/levitsky/pyteomics) or **pymzml** (https://github.com/pymzml/pymzML).
 """
 from .load_adducts import load_adducts
 from .load_from_json import load_from_json
