+++ matchms/importing/__init__.py
@@ -7,8 +7,8 @@ as *.mzML*, *.mzXML*, *.mgf*, or *.msp*. It is also possible to load data from
 *.json* files (tested for json files from GNPS or json files made with matchms).
 Another option is to load spectra based on a unique identifier (USI)
 (:meth:`~matchms.importing.load_from_usi`). For more extensive import options
-we recommend building custom importers using **pyteomics**
-(https://github.com/levitsky/pyteomics) or **pymzml** (https://github.com/pymzml/pymzML).
+we recommend building custom importers using `pyteomics <https://github.com/levitsky/pyteomics>`
+ or `pymzml <https://github.com/pymzml/pymzML>`.
 """
 from .load_adducts import load_adducts
 from .load_from_json import load_from_json
