### 2020-08-17 d44d430
Fixed importing select_by_relative_intensity and intensity_percent check
