### 2020-08-18 ed30d68
add basic code examples
