+++ matchms/exporting/save_as_json.py
@@ -9,6 +9,23 @@ def save_as_json(spectrums: List[Spectrum], filename: str):
 
     :py:attr:`~matchms.Spectrum.losses` of spectrum will not be saved.
 
+    Example:
+
+    .. code-block:: python
+
+        from matchms import Spectrum
+        from matchms.exporting import save_as_json
+
+        # Create dummy spectrum
+        spectrum = Spectrum(mz=numpy.array([100, 200, 300], dtype="float"),
+                            intensities=numpy.array([10, 10, 500], dtype="float"),
+                            metadata={"charge": -1,
+                                      "inchi": '"InChI=1S/C6H12"',
+                                      "precursor_mz": 222.2})
+
+        # Write spectrum to test file
+        save_as_json(spectrum, "test.json")
+
     Parameters
     ----------
     spectrums:
