### 2020-08-18 57db21b
Update matchms/filtering/__init__.py
Co-authored-by: Stefan Verhoeven <s.verhoeven@esciencecenter.nl>
