### 2020-08-18 2fc602c
Update matchms/exporting/save_as_json.py
Co-authored-by: Stefan Verhoeven <s.verhoeven@esciencecenter.nl>
