### 2020-08-18 cfc3743
Update matchms/exporting/save_as_mgf.py
Co-authored-by: Stefan Verhoeven <s.verhoeven@esciencecenter.nl>
