### 2020-08-18 7c1aad4
make code example more similar to other importing functions
