### 2020-08-18 4ccd5c9
make docstring into bullet list
