### 2020-08-18 bd2aed3
minor linting
