### 2020-08-19 5e4795e
add is_symmetric flag
