### 2020-08-19 6dc6165
remove unclear assert
