### 2020-08-19 faced37
add test and bug fix
