### 2020-08-19 b7baae9
add argspec to search for is_symmetric
