### 2020-08-19 matchms/matchms#96 @florian-huber
The remaining parts of this are now in a new issue #107 (and PR #119 ).
