### 2020-08-19 e58ac89
Add tests for remove_peaks_outside_top_k
