### 2020-08-19 ab18c75
Fix argsort usage
