+++ matchms/filtering/remove_peaks_outside_top_k.py
@@ -23,7 +23,7 @@ def remove_peaks_outside_top_k(spectrum: SpectrumType, k: int = 6,
     assert k >= 1, "k must be a positive nonzero integer."
     assert mz_window >= 0, "mz_window must be a positive floating point."
     mzs, intensities = spectrum.peaks
-    top_k = intensities.argsort[::-1][0:k]
+    top_k = intensities.argsort()[::-1][0:k]
     k_ordered_mzs = mzs[top_k]
     indices = [i for i in range(len(mzs)) if i not in top_k]
     new_mzs, new_intensities = mzs, intensities
