### 2020-08-20 #2411 @bittremieux
It looks good for me as well, I just have a few really minor comments:

- Missing reference for mzML on page 1: [10.1074/mcp.R110.000133](http://www.mcponline.org/content/10/1/R110.000133.abstract)
- I would recommend consistent and correct spelling of some terms throughout the manuscript and figures, for example: Inchi/InchiKey -> InChI/InChIKey, Smiles -> SMILES, _m/z_ should be italicized, NumPy, SciPy
