### 2020-08-20 #2411 @florian-huber
Thanks for the comments @bittremieux.
We changed spelling as you suggested (in text and figures).
