### 2020-08-20 3172876
Change using nans as markers
Co-authored-by: Florian Huber <36473328+florian-huber@users.noreply.github.com>
