### 2020-08-20 13a93d3
Fixed comprehension for keep_idx (linting)
