### 2020-08-20 6dd6030
minor docstring consistency linting
