+++ matchms/filtering/remove_peaks_around_precursor_mz.py
@@ -6,13 +6,13 @@ from ..typing import SpectrumType
 def remove_peaks_around_precursor_mz(spectrum: SpectrumType, mz_tolerance: float = 17) -> SpectrumType:
 
     """Remove peaks that are within mz_tolerance (in Da) of
-       the precursor mz, exlcuding the precursor peak
+       the precursor mz, exlcuding the precursor peak.
 
-    Args:
-    -----
-    spectrum:
+    Parameters
+    ----------
+    spectrum
         Input spectrum.
-    mz_tolerance:
+    mz_tolerance
         Tolerance of mz values that are not allowed to lie
         within the precursor mz. Default is 17 Da.
     """
