+++ matchms/filtering/require_minimum_of_high_peaks.py
@@ -8,15 +8,15 @@ def require_minimum_of_high_peaks(spectrum: SpectrumType, no_peaks: int = 5,
     """Returns None if the number of peaks with relative intensity
        above or equal to intensity_percent is less than no_peaks.
 
-    Args:
-    -----
-    spectrum:
+    Parameters
+    ----------
+    spectrum
         Input spectrum.
-    no_peaks:
+    no_peaks
         Minimum number of peaks allowed to have relative intensity
         above intensity_percent. Less peaks will return none.
         Default is 5.
-    intensity_percent:
+    intensity_percent
         Minimum relative intensity (as a percentage between 0-100) for
         peaks that are searched. Default is 2
     """
