### 2020-08-24 464fcdb
add test to check if filter clones spectrum
