+++ tests/test_require_minimum_of_high_peaks.py
@@ -34,3 +34,15 @@ def test_require_minimum_of_high_peaks_intensity_percent_10():
     spectrum = require_minimum_of_high_peaks(spectrum_in, intensity_percent=10)
 
     assert spectrum == spectrum_in, "Expected no changes."
+
+
+def test_if_spectrum_is_cloned():
+    """Test if filter is correctly cloning the input spectrum."""
+    mz = numpy.array([], dtype="float")
+    intensities = numpy.array([], dtype="float")
+    spectrum_in = Spectrum(mz=mz, intensities=intensities)
+
+    spectrum = require_minimum_of_high_peaks(spectrum_in)
+    spectrum.set("testfield", "test")
+
+    assert not spectrum_in.get("testfield"), "Expected input spectrum to remain unchanged."
