### 2020-08-25 matchms/matchms#95 @florian-huber
Final comments were included in the manuscript and reference to mzML was added. I will hence close this issue.
