### 2020-08-25 48aa9fb
Update tests/test_remove_peaks_around_precursor_mz.py
Co-authored-by: Stefan Verhoeven <s.verhoeven@esciencecenter.nl>
