+++ tests/test_remove_peaks_around_precursor_mz.py
@@ -3,8 +3,8 @@ from matchms import Spectrum
 from matchms.filtering import remove_peaks_around_precursor_mz
 
 
-def test_remove_peaks_around_precursor_mz_no_params_or_precursor():
-    """No parameters provided and no precursor mz present."""
+def test_remove_peaks_around_precursor_mz_defaults_no_precursor():
+     """Using defaults and no precursor mz present."""
     mz = numpy.array([10, 20, 30, 40], dtype="float")
     intensities = numpy.array([0, 1, 10, 100], dtype="float")
     spectrum_in = Spectrum(mz=mz, intensities=intensities)
