### 2020-08-25 59147d2
Edit docstring
