+++ matchms/filtering/remove_peaks_outside_top_k.py
@@ -6,8 +6,8 @@ from ..typing import SpectrumType
 def remove_peaks_outside_top_k(spectrum: SpectrumType, k: int = 6,
                                mz_window: float = 50) -> SpectrumType:
 
-    """Keep peaks only if they are in the top k intense peaks
-       within a desired mz (Da) window.
+    """Remove all peaks which are not within *mz_window* of at least one
+       of the *k* highest intensity peaks of the spectrum.
 
     Parameters
     ----------
