### 2020-08-25 afd59aa
Cloned input spectrum and added check for when input spectrum is None
