+++ matchms/filtering/require_minimum_of_high_peaks.py
@@ -2,7 +2,7 @@ from ..typing import SpectrumType
 from .select_by_relative_intensity import select_by_relative_intensity
 
 
-def require_minimum_of_high_peaks(spectrum: SpectrumType, no_peaks: int = 5,
+def require_minimum_of_high_peaks(spectrum_in: SpectrumType, no_peaks: int = 5,
                                   intensity_percent: float = 2.0) -> SpectrumType:
 
     """Returns None if the number of peaks with relative intensity
@@ -20,6 +20,10 @@ def require_minimum_of_high_peaks(spectrum: SpectrumType, no_peaks: int = 5,
         Minimum relative intensity (as a percentage between 0-100) for
         peaks that are searched. Default is 2
     """
+    if spectrum_in is None:
+        return None
+
+    spectrum = spectrum_in.clone()
 
     assert no_peaks >= 1, "no_peaks must be a positive nonzero integer."
     assert 0 <= intensity_percent <= 100, "intensity_percent must be a floating point between 0-100."
