+++ matchms/filtering/require_precursor_below_mz.py
@@ -1,7 +1,7 @@
 from ..typing import SpectrumType
 
 
-def require_precursor_below_mz(spectrum: SpectrumType, max_mz: float = 1000) -> SpectrumType:
+def require_precursor_below_mz(spectrum_in: SpectrumType, max_mz: float = 1000) -> SpectrumType:
 
     """Returns None if the precursor_mz of a spectrum is above
        max_mz.
@@ -15,6 +15,10 @@ def require_precursor_below_mz(spectrum: SpectrumType, max_mz: float = 1000) -> 
         All precursor mz values greater or equal to this
         will return none. Default is 1000.
     """
+    if spectrum_in is None:
+        return None
+
+    spectrum = spectrum_in.clone()
 
     assert max_mz >= 0, "max_mz must be a positive floating point."
     precursor_mz = spectrum.get("precursor_mz")
