### 2020-08-25 b3df608
adapt Scores to new similarity scores classes
