+++ matchms/Scores.py
@@ -1,5 +1,7 @@
 from __future__ import annotations
+import inspect
 import numpy
+from matchms.similarity import ParallelSimilarityFunction
 from matchms.typing import QueriesType
 from matchms.typing import ReferencesType
 from matchms.typing import SimilarityFunction
@@ -49,7 +51,8 @@ class Scores:
         Cosine score between spectrum2 and spectrum3 is 0.14 with 1 matched peaks
         Cosine score between spectrum2 and spectrum4 is 0.61 with 1 matched peaks
     """
-    def __init__(self, references: ReferencesType, queries: QueriesType, similarity_function: SimilarityFunction):
+    def __init__(self, references: ReferencesType, queries: QueriesType,
+                 similarity_function: SimilarityFunction, is_symmetric: bool = False):
         """
 
         Parameters
@@ -60,7 +63,10 @@ class Scores:
             List of query objects
         similarity_function
             Function which accepts a reference + query object and returns a score or tuple of scores
-
+        is_symmetric
+            Set to True when *references* and *queries* are identical (as for instance for an all-vs-all
+            comparison). By using the fact that score[i,j] = score[j,i] the calculation will be about
+            2x faster.
         """
         Scores._validate_input_arguments(references, queries, similarity_function)
 
@@ -69,6 +75,7 @@ class Scores:
         self.references = numpy.asarray(references).reshape(self.n_rows, 1)
         self.queries = numpy.asarray(queries).reshape(1, self.n_cols)
         self.similarity_function = similarity_function
+        self.is_symmetric = is_symmetric
         self._scores = numpy.empty([self.n_rows, self.n_cols], dtype="object")
         self._index = 0
 
@@ -98,27 +105,58 @@ class Scores:
         assert isinstance(queries, (list, tuple, numpy.ndarray)),\
             "Expected input argument 'queries' to be list or tuple or numpy.ndarray."
 
-        assert callable(similarity_function), "Expected input argument 'similarity_function' to be callable."
+    def _chose_parallel_implementation(self):
+        """
+        Chose implementation for calculating the scores based on input dimensions and
+        available score implementations (e.g. parallel or only sequential).
+        """
+        # Check if parallel implementation is NOT provided
+        if not isinstance(self.similarity_function, ParallelSimilarityFunction):
+            return False
 
+        # Check if input is likely to benefit from parallel implementation 
+        if self.n_rows > 1 and self.n_cols > 1:
+            return True
+        return False
+    
     def calculate(self) -> Scores:
+        """
+        Calculate the similarity between all reference objects v all query objects using
+        the most suitable available implementation of the given similarity_function.
+        """
+        if self._chose_parallel_implementation():
+            return self.calculate_parallel()
+        return self.calculate_sequential()
+    
+    def calculate_sequential(self) -> Scores:
         """
         Calculate the similarity between all reference objects v all query objects using a
         naive implementation (i.e. a double for-loop). Similarity functions should expect
         one reference and one query object as its input arguments.
         """
         for i_ref, reference in enumerate(self.references[:self.n_rows, 0]):
-            for i_query, query in enumerate(self.queries[0, :self.n_cols]):
-                self._scores[i_ref][i_query] = self.similarity_function(reference, query)
+            if self.is_symmetric:
+                for i_query, query in enumerate(self.queries[0, i_ref:self.n_cols], start=i_ref):
+                    self._scores[i_ref][i_query] = self.similarity_function.compute_scores(reference, query)
+                    self._scores[i_query][i_ref] = self._scores[i_ref][i_query]
+            else:
+                for i_query, query in enumerate(self.queries[0, :self.n_cols]):
+                    self._scores[i_ref][i_query] = self.similarity_function.compute_scores(reference, query)
[…]
