### 2020-08-25 95f5c7f
add **kwargs to parent class methods
