### 2020-08-25 64fb209
fix text and parent class arguments
