+++ matchms/similarity/AbstractSimilarityFunction.py
@@ -32,7 +32,7 @@ class SequentialSimilarityFunction(BaseSimilarityFunction):
             return s
 
     """
-    def compute_scores(self, reference: SpectrumType, query: SpectrumType, **kwargs) -> float:
+    def compute_scores(self, reference: SpectrumType, query: SpectrumType) -> float:
         raise NotImplementedError
 
 
@@ -70,5 +70,6 @@ class ParallelSimilarityFunction(BaseSimilarityFunction):
             return lens1 * lens2.T
 
     """
-    def compute_scores_parallel(self, reference: List[SpectrumType], query: List[SpectrumType], **kwargs) -> numpy.ndarray:
+    def compute_scores_parallel(self, reference: List[SpectrumType], query: List[SpectrumType],
+                                is_symmetric: bool = None) -> numpy.ndarray:
         raise NotImplementedError
