+++ matchms/Scores.py
@@ -68,6 +68,7 @@ class Scores:
             comparison). By using the fact that score[i,j] = score[j,i] the calculation will be about
             2x faster.
         """
+        #pylint: disable=too-many-arguments
         Scores._validate_input_arguments(references, queries, similarity_function)
 
         self.n_rows = len(references)
