### 2020-08-25 a7ed187
change cosine score
