### 2020-08-25 8e4887e
refactor IntersecMZ score
