### 2020-08-25 cc02263
adapt tests and code examples for Cosine Greedy and INtersetMZ
