+++ tests/test_ParentmassMatch.py
@@ -2,6 +2,8 @@ import numpy
 import pytest
 from matchms import Spectrum
 from matchms.similarity import ParentmassMatch
+from matchms.similarity.ParentmassMatch import parentmass_scores
+from matchms.similarity.ParentmassMatch import parentmass_scores_symmetric
 
 
 def test_parentmass_match():
@@ -51,3 +53,123 @@ def test_parentmass_match_missing_parentmass():
 
     expected_message_part = "Missing parent mass."
     assert expected_message_part in str(msg.value), "Expected particular error message."
+
+
+def test_parentmass_match_array():
+    "Test with array and default tolerance."
+    spectrum_1 = Spectrum(mz=numpy.array([], dtype="float"),
+                          intensities=numpy.array([], dtype="float"),
+                          metadata={"parent_mass": 100.0})
+
+    spectrum_2 = Spectrum(mz=numpy.array([], dtype="float"),
+                          intensities=numpy.array([], dtype="float"),
+                          metadata={"parent_mass": 101.0})
+
+    spectrum_a = Spectrum(mz=numpy.array([], dtype="float"),
+                          intensities=numpy.array([], dtype="float"),
+                          metadata={"parent_mass": 99.0})
+
+    spectrum_b = Spectrum(mz=numpy.array([], dtype="float"),
+                          intensities=numpy.array([], dtype="float"),
+                          metadata={"parent_mass": 98.0})
+
+    similarity_score = ParentmassMatch()
+    scores = similarity_score.compute_scores_parallel([spectrum_1, spectrum_2],
+                                                      [spectrum_a, spectrum_b])
+    assert numpy.all(scores == numpy.array([[False, False],
+                                            [False, False]])), "Expected different scores."
+
+
+def test_parentmass_match_tolerance2():
+    """Test with array and tolerance=2."""
+    spectrum_1 = Spectrum(mz=numpy.array([], dtype="float"),
+                          intensities=numpy.array([], dtype="float"),
+                          metadata={"parent_mass": 100.0})
+
+    spectrum_2 = Spectrum(mz=numpy.array([], dtype="float"),
+                          intensities=numpy.array([], dtype="float"),
+                          metadata={"parent_mass": 101.0})
+
+    spectrum_a = Spectrum(mz=numpy.array([], dtype="float"),
+                          intensities=numpy.array([], dtype="float"),
+                          metadata={"parent_mass": 99.0})
+
+    spectrum_b = Spectrum(mz=numpy.array([], dtype="float"),
+                          intensities=numpy.array([], dtype="float"),
+                          metadata={"parent_mass": 98.0})
+
+    similarity_score = ParentmassMatch(tolerance=2.0)
+    scores = similarity_score.compute_scores_parallel([spectrum_1, spectrum_2],
+                                                      [spectrum_a, spectrum_b])
+    assert numpy.all(scores == numpy.array([[True, True],
+                                            [True, False]])), "Expected different scores."
+
+def test_parentmass_match_symmetric():
+    """Test with array and is_symmetric=True."""
+    spectrum_1 = Spectrum(mz=numpy.array([], dtype="float"),
+                          intensities=numpy.array([], dtype="float"),
+                          metadata={"parent_mass": 100.0})
+
+    spectrum_2 = Spectrum(mz=numpy.array([], dtype="float"),
+                          intensities=numpy.array([], dtype="float"),
+                          metadata={"parent_mass": 101.0})
+
+    spectrum_3 = Spectrum(mz=numpy.array([], dtype="float"),
+                          intensities=numpy.array([], dtype="float"),
+                          metadata={"parent_mass": 99.95})
+
+    spectrum_4 = Spectrum(mz=numpy.array([], dtype="float"),
[…]
