### 2020-08-25 31d2763
fix bug and linting
