### 2020-08-26 16dc3c2
add tests for asserts
