+++ tests/test_require_precursor_below_mz.py
@@ -1,4 +1,5 @@
 import numpy
+import pytest
 from matchms import Spectrum
 from matchms.filtering import require_precursor_below_mz
 
@@ -38,3 +39,27 @@ def test_if_spectrum_is_cloned():
     spectrum.set("testfield", "test")
 
     assert not spectrum_in.get("testfield"), "Expected input spectrum to remain unchanged."
+
+
+def test_require_precursor_below_without_precursor_mz():
+    """Test if correct assert error is raised for missing precursor-mz."""
+    spectrum_in = Spectrum(mz=numpy.array([10, 20, 30, 40], dtype="float"),
+                           intensities=numpy.array([0, 1, 10, 100], dtype="float"),
+                           metadata={})
+
+    with pytest.raises(AssertionError) as msg:
+        _ = require_precursor_below_mz(spectrum_in)
+
+    assert str(msg.value) == "Precursor mz absent.", "Expected different error message."
+
+
+def test_require_precursor_below_with_wrong_precursor_mz():
+    """Test if correct assert error is raised for precursor-mz as string."""
+    spectrum_in = Spectrum(mz=numpy.array([10, 20, 30, 40], dtype="float"),
+                           intensities=numpy.array([0, 1, 10, 100], dtype="float"),
+                           metadata={"precursor_mz": "445.0"})
+
+    with pytest.raises(AssertionError) as msg:
+        _ = require_precursor_below_mz(spectrum_in)
+
+    assert "Expected 'precursor_mz' to be a scalar number." in str(msg.value)
