+++ matchms/filtering/remove_peaks_around_precursor_mz.py
@@ -22,7 +22,7 @@ def remove_peaks_around_precursor_mz(spectrum_in: SpectrumType, mz_tolerance: fl
     spectrum = spectrum_in.clone()
 
     precursor_mz = spectrum.get("precursor_mz", None)
-    assert precursor_mz is not None, "precursor mz absent."
+    assert precursor_mz is not None, "Precursor mz absent."
     assert isinstance(precursor_mz, (float, int)), ("Expected 'precursor_mz' to be a scalar number.",
                                                     "Consider applying 'add_precursor_mz' filter first.")
     assert mz_tolerance >= 0, "mz_tolerance must be a positive scalar."
