### 2020-08-26 1ff378e
add tests for None entry
