+++ tests/test_remove_peaks_outside_top_k.py
@@ -84,3 +84,10 @@ def test_if_spectrum_is_cloned():
     spectrum.set("testfield", "test")
 
     assert not spectrum_in.get("testfield"), "Expected input spectrum to remain unchanged."
+
+
+def test_with_input_none():
+    """Test if input spectrum is None."""
+    spectrum_in = None
+    spectrum = remove_peaks_outside_top_k_mz(spectrum_in)
+    assert spectrum is None
