### 2020-08-26 #2411 @arfon
@bittremieux & @kaibioinfo - if you have a moment to revisit the checklists above to mark off the last few items (which I believe have now been addressed) that would be much appreciated.
