### 2020-08-26 #2411 @arfon
@florian-huber - I've made a couple of minor edits in https://github.com/matchms/matchms/pull/137.

After merging https://github.com/matchms/matchms/pull/137, could you make a new release of this software that includes the changes that have resulted from this review. Then, please make an archive of the software in Zenodo/figshare/other service and update this thread with the DOI of the archive? For the Zenodo/figshare archive, please make sure that:
