### 2020-08-26 753219f
Update to 0.5.2
