### 2020-08-26 #2411 @florian-huber
I merged the minor edits and made a new release (0.5.2) which is now on Zenodo.
The DOI is 10.5281/zenodo.4001991
