### 2020-08-26 33d21d0
add parent class for similarity score
