### 2020-08-26 8f2a7b6
refactor Scores
