+++ matchms/Scores.py
@@ -1,5 +1,8 @@
 from __future__ import annotations
+import inspect
 import numpy
+from matchms.similarity.AbstractSimilarityFunction import \
+    ParallelSimilarityFunction
 from matchms.typing import QueriesType
 from matchms.typing import ReferencesType
 from matchms.typing import SimilarityFunction
@@ -16,7 +19,8 @@ class Scores:
     .. testcode::
 
         import numpy as np
-        from matchms import Scores, Spectrum
+        from matchms import calculate_scores
+        from matchms import Spectrum
         from matchms.similarity import CosineGreedy
 
         spectrum_1 = Spectrum(mz=np.array([100, 150, 200.]),
@@ -34,7 +38,8 @@ class Scores:
         references = [spectrum_1, spectrum_2]
         queries = [spectrum_3, spectrum_4]
 
-        scores = Scores(references, queries, CosineGreedy()).calculate()
+        similarity_measure = CosineGreedy()
+        scores = calculate_scores(references, queries, similarity_measure)
 
         for (reference, query, score, n_matching) in scores:
             print(f"Cosine score between {reference.get('id')} and {query.get('id')}" +
@@ -49,7 +54,8 @@ class Scores:
         Cosine score between spectrum2 and spectrum3 is 0.14 with 1 matched peaks
         Cosine score between spectrum2 and spectrum4 is 0.61 with 1 matched peaks
     """
-    def __init__(self, references: ReferencesType, queries: QueriesType, similarity_function: SimilarityFunction):
+    def __init__(self, references: ReferencesType, queries: QueriesType,
+                 similarity_function: SimilarityFunction, is_symmetric: bool = False):
         """
 
         Parameters
@@ -60,15 +66,20 @@ class Scores:
             List of query objects
         similarity_function
             Function which accepts a reference + query object and returns a score or tuple of scores
-
+        is_symmetric
+            Set to True when *references* and *queries* are identical (as for instance for an all-vs-all
+            comparison). By using the fact that score[i,j] = score[j,i] the calculation will be about
+            2x faster.
         """
-        Scores._validate_input_arguments(references, queries, similarity_function)
+        # pylint: disable=too-many-arguments
+        Scores._validate_input_arguments(references, queries)
 
         self.n_rows = len(references)
         self.n_cols = len(queries)
         self.references = numpy.asarray(references).reshape(self.n_rows, 1)
         self.queries = numpy.asarray(queries).reshape(1, self.n_cols)
         self.similarity_function = similarity_function
+        self.is_symmetric = is_symmetric
         self._scores = numpy.empty([self.n_rows, self.n_cols], dtype="object")
         self._index = 0
 
@@ -91,34 +102,24 @@ class Scores:
         return self._scores.__str__()
 
     @staticmethod
-    def _validate_input_arguments(references, queries, similarity_function):
+    def _validate_input_arguments(references, queries):
         assert isinstance(references, (list, tuple, numpy.ndarray)),\
             "Expected input argument 'references' to be list or tuple or numpy.ndarray."
 
         assert isinstance(queries, (list, tuple, numpy.ndarray)),\
             "Expected input argument 'queries' to be list or tuple or numpy.ndarray."
 
-        assert callable(similarity_function), "Expected input argument 'similarity_function' to be callable."
-
     def calculate(self) -> Scores:
         """
-        Calculate the similarity between all reference objects v all query objects using a
-        naive implementation (i.e. a double for-loop). Similarity functions should expect
-        one reference and one query object as its input arguments.
+        Calculate the similarity between all reference objects v all query objects using
+        the most suitable available implementation of the given similarity_function.
         """
-        for i_ref, reference in enumerate(self.references[:self.n_rows, 0]):
-            for i_query, query in enumerate(self.queries[0, :self.n_cols]):
[…]
