+++ tests/test_ParentmassMatch.py
@@ -2,6 +2,8 @@ import numpy
 import pytest
 from matchms import Spectrum
 from matchms.similarity import ParentmassMatch
+from matchms.similarity.ParentmassMatch import parentmass_scores
+from matchms.similarity.ParentmassMatch import parentmass_scores_symmetric
 
 
 def test_parentmass_match():
@@ -15,7 +17,7 @@ def test_parentmass_match():
                           metadata={"parent_mass": 101.0})
 
     similarity_score = ParentmassMatch()
-    score = similarity_score(spectrum_1, spectrum_2)
+    score = similarity_score.compute_score(spectrum_1, spectrum_2)
     assert not score, "Expected different score."
 
 
@@ -30,7 +32,7 @@ def test_parentmass_match_tolerance2():
                           metadata={"parent_mass": 101.0})
 
     similarity_score = ParentmassMatch(tolerance=2.0)
-    score = similarity_score(spectrum_1, spectrum_2)
+    score = similarity_score.compute_score(spectrum_1, spectrum_2)
     assert score, "Expected different score."
 
 
@@ -47,7 +49,128 @@ def test_parentmass_match_missing_parentmass():
     similarity_score = ParentmassMatch(tolerance=2.0)
 
     with pytest.raises(AssertionError) as msg:
-        _ = similarity_score(spectrum_1, spectrum_2)
+        _ = similarity_score.compute_score(spectrum_1, spectrum_2)
 
     expected_message_part = "Missing parent mass."
     assert expected_message_part in str(msg.value), "Expected particular error message."
+
+
+def test_parentmass_match_array():
+    "Test with array and default tolerance."
+    spectrum_1 = Spectrum(mz=numpy.array([], dtype="float"),
+                          intensities=numpy.array([], dtype="float"),
+                          metadata={"parent_mass": 100.0})
+
+    spectrum_2 = Spectrum(mz=numpy.array([], dtype="float"),
+                          intensities=numpy.array([], dtype="float"),
+                          metadata={"parent_mass": 101.0})
+
+    spectrum_a = Spectrum(mz=numpy.array([], dtype="float"),
+                          intensities=numpy.array([], dtype="float"),
+                          metadata={"parent_mass": 99.0})
+
+    spectrum_b = Spectrum(mz=numpy.array([], dtype="float"),
+                          intensities=numpy.array([], dtype="float"),
+                          metadata={"parent_mass": 98.0})
+
+    similarity_score = ParentmassMatch()
+    scores = similarity_score.compute_score_matrix([spectrum_1, spectrum_2],
+                                                   [spectrum_a, spectrum_b])
+    assert numpy.all(scores == numpy.array([[False, False],
+                                            [False, False]])), "Expected different scores."
+
+
+def test_parentmass_match_tolerance2_array():
+    """Test with array and tolerance=2."""
+    spectrum_1 = Spectrum(mz=numpy.array([], dtype="float"),
+                          intensities=numpy.array([], dtype="float"),
+                          metadata={"parent_mass": 100.0})
+
+    spectrum_2 = Spectrum(mz=numpy.array([], dtype="float"),
+                          intensities=numpy.array([], dtype="float"),
+                          metadata={"parent_mass": 101.0})
+
+    spectrum_a = Spectrum(mz=numpy.array([], dtype="float"),
+                          intensities=numpy.array([], dtype="float"),
+                          metadata={"parent_mass": 99.0})
+
+    spectrum_b = Spectrum(mz=numpy.array([], dtype="float"),
+                          intensities=numpy.array([], dtype="float"),
+                          metadata={"parent_mass": 98.0})
+
+    similarity_score = ParentmassMatch(tolerance=2.0)
+    scores = similarity_score.compute_score_matrix([spectrum_1, spectrum_2],
[…]
