+++ matchms/similarity/ParentmassMatch.py
@@ -1,21 +1,127 @@
+from typing import List
+import numba
+import numpy
+from matchms.similarity.BaseSimilarityFunction import BaseSimilarityFunction
 from matchms.typing import SpectrumType
 
 
-class ParentmassMatch:
-    """Return True if spectrums match in parent mass (within tolerance), and False otherwise."""
+class ParentmassMatch(BaseSimilarityFunction):
+    """Return True if spectrums match in parent mass (within tolerance), and False otherwise.
+
+    Example to calculate scores between 2 spectrums and iterate over the scores
+
+    .. testcode::
+
+        import numpy as np
+        from matchms import calculate_scores
+        from matchms import Spectrum
+        from matchms.similarity import ParentmassMatch
+
+        spectrum_1 = Spectrum(mz=np.array([]),
+                              intensities=np.array([]),
+                              metadata={"id": "1", "parent_mass": 100})
+        spectrum_2 = Spectrum(mz=np.array([]),
+                              intensities=np.array([]),
+                              metadata={"id": "2", "parent_mass": 110})
+        spectrum_3 = Spectrum(mz=np.array([]),
+                              intensities=np.array([]),
+                              metadata={"id": "3", "parent_mass": 103})
+        spectrum_4 = Spectrum(mz=np.array([]),
+                              intensities=np.array([]),
+                              metadata={"id": "4", "parent_mass": 111})
+        references = [spectrum_1, spectrum_2]
+        queries = [spectrum_3, spectrum_4]
+
+        similarity_score = ParentmassMatch(tolerance=5.0)
+        scores = calculate_scores(references, queries, similarity_score)
+
+        for (reference, query, score) in scores:
+            print(f"Parentmass match between {reference.get('id')} and {query.get('id')}" +
+                  f" is {score:.2f}")
+
+    Should output
+
+    .. testoutput::
+
+        Parentmass match between 1 and 3 is 1.00
+        Parentmass match between 1 and 4 is 0.00
+        Parentmass match between 2 and 3 is 0.00
+        Parentmass match between 2 and 4 is 1.00
+
+    """
+    # Set key characteristics as class attributes
+    is_commutative = True
 
     def __init__(self, tolerance: float = 0.1):
         """
         Parameters
         ----------
-        tolerance:
+        tolerance
             Specify tolerance below which two masses are counted as match.
         """
         self.tolerance = tolerance
 
-    def __call__(self, spectrum: SpectrumType, reference_spectrum: SpectrumType) -> bool:
-        """Compare parent masses"""
-        parentmass = spectrum.get("parent_mass")
-        parentmass_ref = reference_spectrum.get("parent_mass")
-        assert parentmass is not None and parentmass_ref is not None, "Missing parent mass."
-        return abs(parentmass-parentmass_ref) <= self.tolerance
+    def compute_score(self, reference: SpectrumType, query: SpectrumType) -> float:
+        """Compare parent masses between reference and query spectrum.
+
+        Parameters
+        ----------
+        reference
+            Single reference spectrum.
+        query
+            Single query spectrum.
+        """
+        parentmass_ref = reference.get("parent_mass")
+        parentmass_query = query.get("parent_mass")
+        assert parentmass_ref is not None and parentmass_query is not None, "Missing parent mass."
+
+        return abs(parentmass_ref - parentmass_query) <= self.tolerance
+
+    def compute_score_matrix(self, references: List[SpectrumType], queries: List[SpectrumType],
+                                is_symmetric: bool = False) -> numpy.ndarray:
+        """Compare parent masses between all references and queries.
+
+        Parameters
+        ----------
+        references
[…]
