### 2020-08-26 31b2869
fix bugs and linting
