### 2020-08-26 5b03c49
refactor CosineGreedy
