### 2020-08-26 b00fd7d
Update matchms/filtering/require_minimum_of_high_peaks.py
Co-authored-by: Stefan Verhoeven <s.verhoeven@esciencecenter.nl>
