### 2020-08-26 cfa6a7c
Update matchms/filtering/remove_peaks_outside_top_k.py
Co-authored-by: Stefan Verhoeven <s.verhoeven@esciencecenter.nl>
