### 2020-08-26 2206321
refactor intersetMZ and add test
