### 2020-08-26 d456fbf
Minor docstring change
