### 2020-08-26 e5f3e0f
Minor docstring change
