### 2020-08-26 49ed479
Minor docstring change
