### 2020-08-26 b602332
Update remove_peaks_outside_top_k.py
