+++ matchms/filtering/remove_peaks_outside_top_k.py
@@ -11,11 +11,11 @@ def remove_peaks_outside_top_k(spectrum_in: SpectrumType, k: int = 6,
 
     Parameters
     ----------
-    spectrum_in
+    spectrum_in:
         Input spectrum.
-    k
+    k:
         The number of most intense peaks to compare to. Default is 6.
-    mz_window
+    mz_window:
         Window of mz values (in Da) that are allowed to lie within
         the top k peaks. Default is 50 Da.
     """
