### 2020-08-26 1744395
remove now redundant implementations
