### 2020-08-26 a38f3b9
remove object dtype from Scores
