### 2020-08-26 6f20e4c
remove object dtype from parent class
