### 2020-08-27 cffea42
add Kia to contributors
