+++ CITATION.cff
@@ -58,6 +58,10 @@ authors:
     family-names: Villanueva Castilla
     given-names: Efrain Manuel
     orcid: "https://orcid.org/0000-0001-7665-3575"
+  -
+    affiliation: "University of Glasgow"
+    family-names: Ashouritaklimi
+    given-names: Kianoosh
 
 cff-version: "1.1.0"
 keywords:
