+++ CITATION.cff
@@ -62,6 +62,7 @@ authors:
     affiliation: "University of Glasgow"
     family-names: Ashouritaklimi
     given-names: Kianoosh
+    orcid: "https://orcid.org/0000-0001-6317-3193"
 
 cff-version: "1.1.0"
 keywords:
