### 2020-08-29 #2411 @arfon
> ```
> INVALID DOIs
>
> - 10.1021/jasms.8b04453 is INVALID
```

Not sure what's going on with that DOI, (perhaps the publisher never registered it) but it's definitely correct for this: https://pubs.acs.org/doi/pdf/10.1021/jasms.8b04453

@florian-huber - this is ready to accept and publish. The EiC on rotation will handle this in the next couple of days.
