### 2020-08-29 #2411 @florian-huber
Thanks a lot @arfon!

And short feedback: as first-time submitter to JOSS I really liked the clear and transparent process, as well as the constructive feedback. Really nice!
